==> code/3_TX_dw_merging.Rmd <==
---
title: "TX Drinking Water Data Merging"
author: "EmmaLi Tsai"
date: "2024-04-10"
output: html_document
---

## Loading packages & functions
```{r}
# packages: 
library(tidyverse)
library(leaflet)
library(sf)
library(aws.s3)

# helper function for intersecting aquifers: 
source("./code/functions/intersect_aquifers.R")
```

## Loading lists of data:  
```{r}
# reading demographic, environmental (gw and sw), water delivery system (wds)
# and financial lists generated in TX_dw_collating: 
demo <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_EPAdemographic_list.RData")

enviro_sw <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_enviro_surface_list.RData")

enviro_gw <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_enviro_ground_list.RData")

wds <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_water_delivery_list.RData")

fin <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_financial_list.RData")

# Grabbing census geographies for data merging: 
census_vars <- c(total_pop = "B01003_001")
census <- tidycensus::get_acs(
  geography = "tract", 
  variables = census_vars, 
  state = "TX", 
  year = 2020,
  geometry = TRUE
)

# Adding counties: 
census_tidy <- census
counties <- unlist(strsplit(census_tidy$NAME, split = ","))
census_tidy$counties <- trimws(counties[grepl("County", counties)])
```
 
## East TX counties
```{r}
# reading counties in east TX from a webscrape: 
east_tx <- read.csv("./data/raw/east_TX_counties.csv") %>%
  select(-X)

# grabbing census geographies: 
east_tx_geom <- census_tidy %>%
  filter(counties %in% paste0(east_tx$county_name, " County")) %>%
  group_by(counties) %>%
  summarize()

# grabbing the larger geography of TX
tx_geom <- census_tidy %>%
  summarize()

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = tx_geom, 
              color = "grey", 
              weight = 1) %>%
  addPolygons(data = east_tx_geom,
              opacity = 9,
              color = "#333c6bff",
              weight = 1)
```

## Merging: surface water (reservoirs, rivers, basins) to pwsids:
```{r}
# merging impaired water bodies with basin information: 
basins <- enviro_sw$major_river_basins %>% 
  left_join(., enviro_sw$impaired_basins, by = "basin_num") %>%
  rename(basin_name_short = basin_name.x, 
         basin_name_long = basin_name.y) %>%
  select(!c("id", "label_no")) %>%
  st_transform(., crs = st_crs(census)) 

# for each reservoir, identify which basin it falls in: 
res <- enviro_sw$reservoirs %>%
  # dropping z coordinates: 
  st_zm(., drop = TRUE) %>%
  # st_as_sf(.) %>%
  st_transform(., crs = st_crs(census)) 
# switch off spherical geometry 
sf_use_s2(FALSE)
res_basin <- st_intersection(res, basins)

# for each river, identify which basin it falls in: 
rivers <- enviro_sw$rivers %>%
  mutate(name = toupper(name)) %>%
  st_transform(., crs = st_crs(census)) %>%
  rename(river_name = name)
riv_basin <- st_intersection(rivers, basins) %>%
  select(river_name:geometry) %>%
  select(-shape_len) %>%
  rename(res_name = river_name) %>%
  mutate(type = "River", 
         status = "Existing") %>%
  unique()

# combining rivers and reservoirs: 
riv_res_basin <- rbind(riv_basin, res_basin) 

# matching reservoir, basins, and rivers with surface water intakes: 
sw <- enviro_sw$surface_water_intakes %>%
  select(-objectid) 
# rectifying differences in horizontal datum - some are in NAD83 and others
# are NAD27: 
nad_27 <- sw %>%
  filter(horz_datum == "NAD27") %>%
  st_as_sf(., coords = c("long_dd", "lat_dd"), 
           crs = st_crs(census)) 
nad_83 <- sw %>%
  filter(horz_datum != "NAD27") %>%
  st_as_sf(., coords = c("long_dd", "lat_dd"), 
           crs = st_crs(census))
# adding these back together 
sw_corrected <- rbind(nad_27, nad_83) %>%
  select(-horz_datum) %>%
  mutate(lon = sf::st_coordinates(.)[,1],
         lat = sf::st_coordinates(.)[,2]) %>%
  as.data.frame() %>%
  select(-geometry)

# matching reservoir, basins, and river with surface water intakes using 
 # a name match: 
riv_res_basin_df <- as.data.frame(riv_res_basin) %>%
  select(-geometry) %>%
  unique()
sw_basins <- merge(riv_res_basin_df, sw_corrected,
                   by.x = "res_name", by.y = "waterbody",
                   all.y = TRUE) %>%
  # removing reservoirs that are not a water supply
  filter(type == "Water Supply" | is.na(type) | type == "River") %>%
  unique()

# converting sw points to sf object: 
points_sf <- sw_basins %>%
  as.data.frame() %>%
  filter(!(is.na(lat))) %>%
  st_as_sf(., coords = c("lon", "lat"),
           crs = st_crs(census)) %>%
  # retain coordinates:
  mutate(lon = sf::st_coordinates(.)[,1],
         lat = sf::st_coordinates(.)[,2])

# grabbing entries that are missing from the original merge and will require 
# matching using st_intersection:
no_match <- sw_basins %>%
  filter(is.na(type)) 

# intersection with a 0.01dd buffer to find nearby water bodies that likely 
# match to the sw intake: 
# turning spherical geometry back on:
sf_use_s2(TRUE) 
missing_sw <- points_sf %>%
  filter(res_name %in% no_match$res_name) %>%
  st_buffer(., dist = 0.01)

sw_resbasin <- st_intersection(missing_sw, riv_res_basin) %>%
  select(-c(basin_num:status))
# resolved 49 sw intakes 

# adding in basin information for those that don't match with any water body: 
no_waterbody <- missing_sw %>%
  filter(!(paste0(res_name, pwsid, sys_name)) %in% 
           paste0(sw_resbasin$res_name, sw_resbasin$pwsid, 
                  sw_resbasin$sys_name))
no_waterbody_basin <- st_intersection(no_waterbody, basins) %>%
  select(-c(basin_num:status))

# why are we missing two? 
# One of them is MILLWOOD LAKE which is outside of TX
# and another is IBWC FALCON VILLAGE in the Falcon Reservoir - will need to be fixed by hand at the end. 
falcon <- no_waterbody_basin %>%
  filter(res_name == "FALCON RESERVOIR")

missing_falcon <- no_waterbody %>%
  filter(sys_name == "IBWC FALCON VILLAGE") %>%
  mutate(basin_num.1 = falcon$basin_num, 
         basin_name_short.1 = falcon$basin_name_short, 
         basin_name_long.1 = falcon$basin_name_long, 
         length_impaired.1 = falcon$length_impaired, 
         length_total.1 = falcon$length_total, 
         pct_impaired.1 = falcon$pct_impaired)

missing_falcon <- missing_falcon[,names(no_waterbody_basin)]
filled_intake_data <- rbind(no_waterbody_basin, missing_falcon)
filled_intake_data$res_name.1 <- "no matching wtrbdy"
filled_intake_data$status.1 <- "no matching wtrbdy"
filled_intake_data$type.1 <- "no matching wtrbdy"

# combining back with missing surface water from intersections: 
filled_intake_data <- filled_intake_data[, names(sw_resbasin)]
all_filled_intakes <- rbind(sw_resbasin, filled_intake_data) %>%
  as.data.frame() %>%
  select(-geometry)

# splitting and recombining back into the larger dataset:  
match <- sw_basins %>%
  filter(!is.na(type)) %>%
  mutate(alt_res_name = "NA")

all_filled_intakes <- all_filled_intakes %>%
  rename(alt_res_name = res_name.1)
names(all_filled_intakes) <- gsub(".[[:digit:]]", "", names(all_filled_intakes)) 

all_sw_intakes <- match[,names(all_filled_intakes)] 
all_sw_intakes <- rbind(all_filled_intakes, all_sw_intakes) 

# doing a last st_intersect to match sw intakes to the 
# correct basin, as some rivers span multiple: 
all_sw <- all_sw_intakes %>%
  select(-c(basin_num:status)) %>%
  st_as_sf(., coords = c("lon", "lat"), 
           crs = st_crs(census)) %>%
  filter(sys_name != "IBWC FALCON VILLAGE")
all_intakes_basin_fixed <- st_intersection(all_sw, basins)
# adding falcon back, since it doesn't automatically match to 
# Rio Grande: 
falcon <- all_sw_intakes %>%
  filter(sys_name == "IBWC FALCON VILLAGE") %>%
  st_as_sf(., coords = c("lon", "lat"), 
           crs = st_crs(census)) %>%
  mutate(lon = sf::st_coordinates(.)[,1],
         lat = sf::st_coordinates(.)[,2])
falcon <- falcon[, names(all_intakes_basin_fixed)]
all_intakes_basin_fixed <- rbind(all_intakes_basin_fixed, falcon)

# final tidying of dataset: 
all_sw_intakes_final <- all_intakes_basin_fixed %>%
  unique() %>%
  mutate(alt_res_name = 
           case_when(alt_res_name == "NA" ~ res_name, 
                     TRUE ~ alt_res_name))

# saving: 
st_write(all_sw_intakes_final, "./data/clean/pwsid_swaterbodies.geojson")
# writing to aws:
put_object(
  file = "./data/clean/pwsid_swaterbodies.geojson",
  object = "/state-drinking-water/TX/clean/pwsid_swaterbodies.geojson",
  bucket = "tech-team-data"
)

all_sw_intakes <- st_read("./data/clean/pwsid_swaterbodies.geojson")

# plotting: 
wtrbdy_pal <- colorFactor(
  palette = viridis::plasma(23),
  domain = basins$basin_name_short)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = basins,
              opacity = 9,
              color = ~wtrbdy_pal(basin_name_short),
              weight = 1,
              label = paste0("basin: ", basins$basin_name_short, "; % impaired: ", round(basins$pct_impaired, 2))) %>%
  addLegend("bottomright",
            pal = wtrbdy_pal,
            values = basins$basin_name_short,
            title = "Basin",
            opacity = 1) %>%
  addPolygons(data = res_basin,
              opacity = 9,
              color = ~wtrbdy_pal(basin_name_short),
              weight = 1,
              label = paste0("reservoir: ", res_basin$res_name, "; basin: ",
                             res_basin$basin_name_short)) %>%
  addCircleMarkers(data = all_sw_intakes, 
                   fill = ~wtrbdy_pal(basin_name_short), 
                   color = ~wtrbdy_pal(basin_name_short), 
                   weight = 3, 
                   opacity = 9, 
                   radius = 3, 
                   label = paste0("pwsid: ", all_sw_intakes$pwsid, "; waterbody: ", 
                                  all_sw_intakes$res_name, 
                                  "; basin name: ", 
                                  all_sw_intakes$basin_name_short)) %>%
  addPolylines(data = riv_basin,
               opacity = 9,
               color = ~wtrbdy_pal(basin_name_short),
               weight = 2,
               label = paste0("river: ", riv_basin$res_name))
```

## Merging: groundwater (aquifers, wells) to counties:
```{r}
# loading in minor and major aquifer boundaries: 
minor_aquifers <- enviro_gw$minor_aquifers %>%
  st_transform(., crs = st_crs(census))

major_aquifers <- enviro_gw$major_aquifers %>%
  st_transform(., crs = st_crs(census))

# loading in well locations and ONLY looking at wells where the water primarily 
# goes to a public supply: 
well_twdb <- enviro_gw$well_locations %>%
  filter(primary_wat == "Public Supply") %>%
  st_transform(., crs = st_crs(census)) %>%
  janitor::clean_names()

# plotting to investigate overlap between the minor and major aquifers: 
aq_pal <- colorFactor(
  palette = viridis::plasma(22),
  domain = minor_aquifers$aqu_name)

maj_aq_pal <- colorFactor(
  palette = viridis::mako(9),
  domain = major_aquifers$aq_name)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = major_aquifers,
              opacity = 9,
              color = ~maj_aq_pal(aq_name),
              weight = 1,
              label = paste0("major aq: ", major_aquifers$aq_name)) %>%
  addPolygons(data = minor_aquifers,
              opacity = 9,
              color = ~aq_pal(aqu_name),
              weight = 1,
              label = paste0("minor aq: ", minor_aquifers$aqu_name)) %>%
  addLegend("bottomright",
            pal = aq_pal,
            values = minor_aquifers$aqu_name,
            title = "Minor Aquifers",
            opacity = 1) %>%
  addLegend("bottomright",
            pal = maj_aq_pal,
            values = major_aquifers$aq_name,
            title = "Major Aquifers",
            opacity = 1) 


# running intersections of wells to aquifer boundaries: 
well_major_aquifers <- intersect_aquifers(wells = well_twdb, 
                                          aquifers = major_aquifers, 
                                          aquifer_name_col = "aq_name")

well_minor_aquifers <- intersect_aquifers(wells = well_twdb, 
                                          aquifers = minor_aquifers, 
                                          aquifer_name_col = "aqu_name")

# combining intersected points into a single data frame
wells_major <- well_major_aquifers %>%
  select(-c("aquifer", "perimeter")) %>%
  mutate(aquifer_type = "major") %>%
  rename(aquifer_area = area, 
         aquifer_name = aq_name) %>%
  relocate(aquifer_area, .after = aquifer_name) %>%
  st_transform(., crs = st_crs(census))

wells_minor <- well_minor_aquifers %>%
  select(-c("aquifer", "minor_area")) %>%
  mutate(aquifer_type = "minor") %>%
  rename(aquifer_area = area_aq, 
         aquifer_name = aqu_name) %>%
  st_transform(., crs = st_crs(census))

# adding an aquifer code to reorg the data frame to wide format with each 
# row being a single well number: 
all_well_data <- rbind(wells_major, wells_minor) %>%
  mutate(aquifer_type_code = paste0(aquifer_name, "-", aquifer_type))

# splitting data by  unique well number: 
grouped_wells_aquifer <- split(all_well_data$aquifer_type_code,
                               all_well_data$state_well_n)
unique_wells <- lapply(grouped_wells_aquifer, unique)
# what the maximum number of aquifers a well intersects with (5?!)
max(lengths(unique_wells)) 

# transforming to data frame
well_aquifers <- plyr::ldply(unique_wells, rbind)
names(well_aquifers) <- c("state_well_n", "aquifer_one", "aquifer_two", 
                          "aquifer_three", "aquifer_four", "aquifer_five")

# merging back in with well data: 
well_final <- merge(well_twdb, 
                    well_aquifers, all.x = TRUE) 
# how many didn't intersect with any aquifer boundary?
sum(is.na(well_final$aquifer_one)) # 261 

# writing: 
st_write(well_final, "./data/clean/well_county_aquifers.geojson")
# writing to aws:
put_object(
  file = "./data/clean/well_county_aquifers.geojson",
  object = "/state-drinking-water/TX/clean/well_county_aquifers.geojson",
  bucket = "tech-team-data"
)
well_final <- st_read("./data/clean/well_county_aquifers.geojson")

# plotting wells points and aquifers together:  
well_pal <- colorFactor(
  palette = viridis::mako(20),
  domain = well_final$aquifer_one)

maj_aq_pal <- colorFactor(
  palette = viridis::mako(9),
  domain = major_aquifers$aq_name)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = major_aquifers,
              opacity = 9,
              color = ~maj_aq_pal(aq_name),
              weight = 1,
              label = paste0("major aq: ", major_aquifers$aq_name)) %>%
  addPolygons(data = minor_aquifers,
              opacity = 9,
              color = ~aq_pal(aqu_name),
              weight = 1,
              label = paste0("minor aq: ", minor_aquifers$aqu_name)) %>%
  addLegend("bottomright",
            pal = aq_pal,
            values = minor_aquifers$aqu_name,
            title = "Minor Aquifers",
            opacity = 1) %>%
  addLegend("bottomright",
            pal = maj_aq_pal,
            values = major_aquifers$aq_name,
            title = "Major Aquifers",
            opacity = 1) %>%
  addCircleMarkers(data = well_final,
                   opacity = 9,
                   radius = 0.2,
                   color = ~well_pal(aquifer_one),
                   weight = 1,
                   label = paste0("well #:",
                                  well_final$state_well_n,
                                  "; county: ",
                                  well_final$county_name,
                                  "; aquifer one: ",
                                  well_final$aquifer_one,
                                  "; aquifer two: ",
                                  well_final$aquifer_two,
                                  "; aquifer three: ",
                                  well_final$aquifer_three,
                                  "; aquifer four: ",
                                  well_final$aquifer_four,
                                  "; aquifer five: ",
                                  well_final$aquifer_five))
```

## Merging keys: 
```{r}
# key of sw intake, pwsid, water body, and basin name: 
sw_intakes <- aws.s3::s3read_using(st_read, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/pwsid_swaterbodies.geojson")

# pwsid, waterbody name, and basin name: 
pwsid_sw_merge <- sw_intakes %>%
  as.data.frame() %>%
  select(pwsid, sys_name, basin_name_short, res_name, alt_res_name) %>%
  unique() %>%
  # filtering NAs, from when a waterbody exists but does not supply water 
  # to a public water system: 
  filter(!is.na(pwsid))
# 355 pwsids


# key of wells, counties, and all intersecting aquifers: 
gw_wells <- aws.s3::s3read_using(st_read, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/well_county_aquifers.geojson")
# well number, county, and aquifers: 
well_gw_merge <- gw_wells %>%
  as.data.frame() %>%
  select(state_well_n, county_name, aquifer_one:aquifer_five) %>%
  unique()
# 12997 wells


# key for variables included in most analyses, such as MHI, 
# %POC, and violations (5yr and 10yr) - paperwork and health-based
census_keys <- demo$census %>%
  select("pwsid", "estimate_mhi", "estimate_poc_alone_per")

# grabbing sdiws violations, pwsids, and year
sdwis <- wds$sdwis
pwsids <- unique(census_keys$pwsid)
year <- year(Sys.Date())

# calculating number of violations over the past 5 years: 
viol_5yr <- sdwis %>% 
  filter(pwsid %in% pwsids) %>%
  filter(viol_year >= (year - 5)) %>%
  group_by(pwsid, is_health_based_ind) %>%
  summarize(total_viol = length(unique(violation_id))) %>%
  pivot_wider(names_from = is_health_based_ind, values_from = total_viol) %>%
  rename(paperwork_violations_5yr = N, 
         healthbased_violations_5yr = Y) %>%
  mutate_at(c("paperwork_violations_5yr", "healthbased_violations_5yr"), 
            ~replace_na(.,0)) %>%
  mutate(total_violations_5yr = sum(paperwork_violations_5yr + healthbased_violations_5yr))

# calculating number of violations over the past 10 years: 
viol_10yr <- sdwis %>% 
  filter(pwsid %in% pwsids) %>%
  filter(viol_year >= (year - 10)) %>%
  group_by(pwsid, is_health_based_ind) %>%
  summarize(total_viol = length(unique(violation_id))) %>%
  pivot_wider(names_from = is_health_based_ind, values_from = total_viol) %>%
  rename(paperwork_violations_10yr = N, 
         healthbased_violations_10yr = Y) %>%
  mutate_at(c("paperwork_violations_10yr", "healthbased_violations_10yr"), 
            ~replace_na(.,0)) %>%
  mutate(total_violations_10yr = sum(paperwork_violations_10yr + healthbased_violations_10yr))

# combining into one data frame: 
violation_summary <-merge(viol_10yr, viol_5yr, by = "pwsid", all.x = TRUE) %>%
  # mutating since some pwsids might not have had violations over the past 
  # 5 years
  mutate_at(c("paperwork_violations_5yr", "healthbased_violations_5yr", 
              "total_violations_5yr"), 
            ~replace_na(.,0)) 

# merging with census keys
analysis_keys <- merge(census_keys, violation_summary, by = "pwsid", 
                       all = TRUE)
# for all violation columns - any NAs actually represent zero 
analysis_keys <- analysis_keys %>%
  mutate(across(paperwork_violations_10yr:total_violations_5yr, ~replace_na(.,0)))
# NOTE: NAs in census columns indicate they're missing from the tract crosswalk

# adding this to a larger list: 
TX_merging_keys <- list()
TX_merging_keys[[1]] <- analysis_keys
TX_merging_keys[[2]] <- pwsid_sw_merge
TX_merging_keys[[3]] <- well_gw_merge

names(TX_merging_keys) <- c("analysis_keys", "pwsid_sw", "well_gw_county")
saveRDS(TX_merging_keys, file="./data/clean/TX_EPAmerging_keys_list.RData")
put_object(
  file = "./data/clean/TX_EPAmerging_keys_list.RData",
  object = "/state-drinking-water/TX/clean/TX_EPAmerging_keys_list.RData",
  bucket = "tech-team-data",
)
```
