==> code/4_TX_dw_analysis_gov.Rmd <==
---
title: "TX Drinking Water Data Analysis - Hydrology"
author: "Gabe Watson"
date: "2024-02-27"
laste_updated: "2024-03-20"
output: html_document
---

## Loading packages

```{r}
library(tidyverse)
library(leaflet)
library(sf)
library(aws.s3)
library(vtable)
library(ggplot2)
library(corrplot)
library(forcats)

```

## Loading data lists

```{r}
# reading demographic, environmental (gw and sw), water delivery system (wds)
# and financial lists generated in TX_dw_collating: 
demo <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_demographic_list.RData")
enviro_sw <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_enviro_surface_list.RData")
enviro_gw <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_enviro_ground_list.RData")
wds <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_water_delivery_list.RData")
fin <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_financial_list.RData")
keys <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_merging_keys_list.RData")
crosswalk <- aws.s3::s3read_using(read.csv, 
                                  object = "s3://tech-team-data/pws_crosswalk/pws_census_tract_weighted_crosswalk.csv")
```

## Do residents trust (or drink) their tap water?

# going to pull CCR violations and other metrics of public engagement

```{r}


```

## Who owns the water system?

```{r}
owner_type_code <- c("F","L","M","N","P","S")
## from sdwis
owner_type_description <-c("Federal Government","Local Government","Public/Private","Native American","Private","State Government")
owner_type_merge <- data.frame(owner_type_code,owner_type_description)


## Filtering to only TX systems
ownership <- demo$census %>%
             data.frame()%>%
             left_join(owner_type_merge)%>%
             left_join(data.frame(keys$analysis_keys))%>%
             filter(primacy_agency_code == "TX")

## Ownership Summary
##   Ownership Type Count + %
##   Population Served Count + %
##   HB + PW by Ownership + 
##   MHI By Ownership Mean
##   POC By Ownership Mean
##   Potential Followup Questions 
ownershipsummary <- ownership %>%
                    group_by(owner_type_description, .drop=FALSE)%>%
                    filter(east_tx_flag == "yes")%>%
                    add_tally()%>%
                    summarise(estimate_total_pop = sum(estimate_total_pop, na.rm = TRUE), 
                              n = mean(n, na.rm = TRUE), 
                              paperwork_violations_5yr = mean(paperwork_violations_5yr, na.rm = TRUE), 
                              healthbased_violations_5yr = mean(healthbased_violations_5yr, na.rm = TRUE), 
                              paperwork_violations_10yr = mean(paperwork_violations_10yr, na.rm = TRUE), 
                              healthbased_violations_10yr = mean(healthbased_violations_10yr, na.rm = TRUE), 
                              estimate_poc_alone_per = mean(estimate_poc_alone_per, na.rm = TRUE),
                              estimate_mhi= mean(estimate_mhi, na.rm = TRUE))%>%
                    rename(owner_type_count = n)%>% 
                    mutate(percent_pop = estimate_total_pop / sum(estimate_total_pop))%>%
                    mutate(percent_systems = owner_type_count / sum(owner_type_count))

## Quick Summary ##
## For all of TEXAS! 
## 61% of systems are local public, but serve 91% of population, private systems comprise 38% of systems and 9% of population.
## Public systems are slightly lower income (77k priv vs 74k), more poc (26% vs 21%), and more hb violations (1.6 vs 1.3), BUT 
## Private systems have 17.2 pwork violations vs 5.7 for public...
## 

## ownership pop 
ggplot(data=ownershipsummary, aes(x=owner_type_description, y=percent_pop*100, fill=owner_type_description)) + 
  geom_bar(stat = "identity", color="black")+
  geom_text(aes(label=round(percent_pop*100, digits = 2)), vjust=-0.5)+
  ylab("percent_pop")+
  theme_minimal()+
  ggtitle("East Texas CWS Ownership Types by Percent of CWS Population")
  
## ownership hb
ggplot(data=ownershipsummary, aes(x=owner_type_description, y=healthbased_violations_10yr, fill=owner_type_description)) + 
  geom_bar(stat = "identity", color="black")+
  geom_text(aes(label=round(healthbased_violations_10yr, digits = 2)), vjust=-0.5)+
  theme_minimal()+
  ggtitle("East Texas CWS Ownership Types by Health Based Violations - 10 years")

## ownership pw 
ggplot(data=ownershipsummary, aes(x=owner_type_description, y=paperwork_violations_10yr, fill=owner_type_description)) + 
  geom_bar(stat = "identity", color="black")+
  geom_text(aes(label=round(paperwork_violations_10yr, digits = 2)), vjust=-0.5)+
  theme_minimal()+
  ggtitle("East Texas CWS Ownership Types by Paperwork Violations - 10 years")

```

## What is the size of the water system?

```{r}
## Number of connections 
## Pop density
## Pop count
## Area 
## What is the size of the water system?

## Exploring the data real quick 
## 100 groupings
## Also population groupings 
## cats
#Very Small: 1,000 or less*
#Small: 1,001 - 3,300 
#Medium: 3,301 - 10,000 
#Large: 10,001 - 100,000
#Very Large: Greater than 100,000

system <- demo$census %>%
          data.frame()%>%
          filter(east_tx_flag == "yes")%>%
          left_join(keys$analysis_keys)%>%
          mutate(pop_cats = ntile(estimate_total_pop, 100))%>%
          mutate(area_cats = ntile(area_miles, 100))%>%
          mutate(pop_density_cats = ntile(pop_density, 100))%>%
                              mutate(pop_catagories = ifelse(estimate_total_pop > 0, "Very Small 0-1k", "Very Small (0 - 1,000)"))%>%
                              mutate(pop_catagories = ifelse(estimate_total_pop >= 1001, "Small 1k-3.3k", pop_catagories))%>%
                              mutate(pop_catagories = ifelse(estimate_total_pop >= 3301, "Medium 3.3k-10k", pop_catagories))%>%
                              mutate(pop_catagories = ifelse(estimate_total_pop >= 10001, "Large 10k-100k", pop_catagories))%>%
                              mutate(pop_catagories = ifelse(estimate_total_pop > 100000, "Very Large 100k+", pop_catagories))%>%
                              mutate(pop_catagories = ifelse(is.na(estimate_total_pop), "Tribal", pop_catagories))

# summary stat table 
#st(system)
# correlation matrix
M = cor(system %>% select(c(
                            healthbased_violations_5yr,paperwork_violations_5yr, area_miles, 
                            estimate_white_alone_per:estimate_poc_alone_per)), use="pairwise.complete.obs")
corrplot(M, method = 'number') 
## Nothing telling here, shifting approach to percentiles and groupings ## 

## Summarizing by tiles
pop_cats <- system %>%
                       group_by(pop_cats)%>%
                       add_tally()%>%
                         summarise(pop_total = sum(estimate_total_pop, na.rm = TRUE), 
                              n = mean(n, na.rm = TRUE), 
                              paperwork_violations_10yr = mean(paperwork_violations_10yr, na.rm = TRUE), 
                              healthbased_violations_10yr = mean(healthbased_violations_10yr, na.rm = TRUE), 
                              estimate_poc_alone_per = mean(estimate_poc_alone_per, na.rm = TRUE),
                              estimate_mhi= mean(estimate_mhi, na.rm = TRUE), 
                              area_mean = mean(area_miles, na.rm = TRUE),
                              pop_mean = mean(estimate_total_pop))
hist(pop_cats$pop_mean)
p <- cor(pop_cats, use="pairwise.complete.obs")
corrplot(p, method = 'number')

area_cats <- system %>%
                       group_by(area_cats)%>%
                       add_tally()%>%
                         summarise(estimate_total_pop_sum = sum(estimate_total_pop, na.rm = TRUE), 
                              n = mean(n, na.rm = TRUE), 
                              paperwork_violations_10yr = mean(paperwork_violations_10yr, na.rm = TRUE), 
                              healthbased_violations_10yr = mean(healthbased_violations_10yr, na.rm = TRUE), 
                              estimate_poc_alone_per = mean(estimate_poc_alone_per, na.rm = TRUE),
                              estimate_mhi= mean(estimate_mhi, na.rm = TRUE), 
                              pop_mean = mean(estimate_total_pop, na.rm = TRUE),
                              area_mean = mean(area_miles, na.rm = TRUE))

a <- cor(area_cats, use="pairwise.complete.obs")
corrplot(a, method = 'number')

density_cats <- system %>%
                       group_by(pop_density_cats)%>%
                       add_tally()%>%
                         summarise(estimate_total_pop_sum = sum(estimate_total_pop, na.rm = TRUE), 
                              n = mean(n, na.rm = TRUE), 
                              paperwork_violations_10yr = mean(paperwork_violations_10yr, na.rm = TRUE), 
                              healthbased_violations_10yr = mean(healthbased_violations_10yr, na.rm = TRUE), 
                              estimate_poc_alone_per = mean(estimate_poc_alone_per, na.rm = TRUE),
                              estimate_mhi= mean(estimate_mhi, na.rm = TRUE), 
                              pop_mean = mean(estimate_total_pop, na.rm = TRUE),
                              area_mean = mean(area_miles, na.rm = TRUE))

d <- cor(density_cats, use="pairwise.complete.obs")
corrplot(d, method = 'number')

ggplot(system, aes(x = estimate_total_pop, color = "red")) + 
  geom_histogram()+
  xlim(0,250000)+
  ylim(0,300)+
  theme_minimal()

### There's some interesting things happening year, but need more time to dig into it - quick thoughts!
## -  As population density and population goes down, health based violations and especially paperwork violations go up
## - Seeing some pattern where medium density, pop, and size have the fewest hb violations. 

## pop_density x hb x mhi 
ggplot(density_cats, aes(x=pop_density_cats,y= healthbased_violations_10yr,fill = estimate_mhi))+
  geom_point(shape = 21, color = "black", size = 3)+
   scale_fill_gradient(low = "yellow", high = "red", na.value = NA)+
  theme_classic()

## pop_density x hb x poc 
ggplot(density_cats, aes(x=pop_density_cats,y= healthbased_violations_10yr,fill = estimate_poc_alone_per))+
  geom_point(shape = 21, color = "black", size = 3)+
   scale_fill_gradient(low = "yellow", high = "red", na.value = NA)+
  theme_classic()

## pop x hb x mhi 
ggplot(pop_cats, aes(x=pop_cats,y= healthbased_violations_10yr, fill = estimate_mhi))+
   geom_point(shape = 21, color = "black", size = 3)+
   scale_fill_gradient(low = "yellow", high = "red", na.value = NA)+
  theme_classic()

## area x hb x mhi 
ggplot(area_cats, aes(x=area_cats,y= healthbased_violations_10yr, fill = estimate_mhi))+
  geom_point(shape = 21, color = "black", size = 3)+
   scale_fill_gradient(low = "yellow", high = "red", na.value = NA)+
  theme_classic()

## area x hb x poc 
ggplot(area_cats, aes(x=area_cats,y= healthbased_violations_10yr, fill = estimate_poc_alone_per))+
  geom_point(shape = 21, color = "black", size = 3)+
   scale_fill_gradient(low = "yellow", high = "red", na.value = NA)+
  theme_classic()

## density x pw x poc 
ggplot(density_cats, aes(x=pop_density_cats,y= paperwork_violations_10yr, fill = estimate_mhi))+
  geom_point(shape = 21, color = "black", size = 3)+
   scale_fill_gradient(low = "yellow", high = "red", na.value = NA)+
  theme_classic()

## pop x pw x mhi 
ggplot(pop_cats, aes(x=pop_cats,y= paperwork_violations_10yr, fill = estimate_mhi))+
    geom_point(shape = 21, color = "black", size = 3)+
   scale_fill_gradient(low = "yellow", high = "red", na.value = NA)+
  theme_classic()

## area x pw x mhi 
ggplot(area_cats, aes(x=area_cats,y= paperwork_violations_10yr, fill = estimate_mhi))+
   geom_point(shape = 21, color = "black", size = 3)+
   scale_fill_gradient(low = "yellow", high = "red", na.value = NA)+
   theme_classic()

## Summary Stats ##
summary_pop <- system %>%
               group_by(pop_catagories)%>%
               summarise(estimate_total_pop_sum = sum(estimate_total_pop, na.rm = TRUE), 
               estimate_mhi= mean(estimate_mhi, na.rm = TRUE),
               paperwork_violations_10yr = mean(paperwork_violations_10yr, na.rm = TRUE), 
               healthbased_violations_5yr = mean(healthbased_violations_5yr, na.rm = TRUE),
               healthbased_violations_10yr = mean(healthbased_violations_10yr, na.rm = TRUE))%>%
               mutate(percent_pop = estimate_total_pop_sum / sum(system$estimate_total_pop, na.rm = TRUE))%>%
               mutate(pop_catagories = fct_reorder(pop_catagories, desc(paperwork_violations_10yr)))

## system size 
ggplot(data=summary_pop, aes(x=pop_catagories, y=percent_pop, fill=pop_catagories)) +
  geom_bar(stat = "identity", color="black")+
  geom_text(aes(label=round(percent_pop*100, digits = 2)), vjust=-0.5)+
  theme_minimal()+
  ggtitle("East Texas CWS System Size by Percent of CWS Population")
  
## system size hb
ggplot(data=summary_pop, aes(x=pop_catagories, y=healthbased_violations_10yr, fill=pop_catagories)) + 
  geom_bar(stat = "identity", color="black")+
  geom_text(aes(label=round(healthbased_violations_10yr, digits = 2)), vjust=-0.5)+
  theme_minimal()+
  ggtitle("East Texas CWS System Size by Health Based Violations - 10 years")

## system size pw 
ggplot(data=summary_pop, aes(x=pop_catagories, y=paperwork_violations_10yr, fill=pop_catagories)) + 
  geom_bar(stat = "identity", color="black")+
  geom_text(aes(label=round(paperwork_violations_10yr, digits = 2)), vjust=-0.5)+
  theme_minimal()+
  ggtitle("East Texas CWS System Size by Paperwork Violations - 10 years")


```

## Spatial Analylsis
## Note: This analysis has been sunsetted in favor of EmmaLi's 

```{r}
## Data Prep 
systems_spatial <- demo$census %>%
                   left_join(data.frame(keys$analysis_keys))



tx_counties <- st_read("data/raw/texas_counties.geojson")%>%
               rename(county_served = name)%>%
               mutate(county_served = toupper(county_served))%>%
               select(county_served)

## County Summary 
county_summary <- systems_spatial %>%
                  data.frame()%>%
                  group_by(county_served)%>%
                  summarise(healthbased_violations_10yr = mean(healthbased_violations_10yr, na.rm = TRUE),
                            paperwork_violations_10yr = mean(paperwork_violations_10yr, na.rm = TRUE))%>%
                  left_join(east_tx)%>%
                  left_join(tx_counties)%>%
                  st_as_sf()

## County Map
# color pal
countypal <- colorNumeric(
  palette = "Reds",
  domain = county_summary$healthbased_violations_5yr)

  leaflet()%>%
  addPolygons(data = county_summary, fillColor = ~countypal(healthbased_violations_10yr), color = "black", weight = 1.5)
                  
## Ok none of the counties we're curious about get highlighted. How about a map of utilities - colored by HB
  systempalhb <- colorNumeric(
  palette = "Reds",
  domain = systems_spatial %>% filter(county_served %in% east_tx$county_served) %>% pull(healthbased_violations_10yr))
  
  leaflet()%>%
  addCircleMarkers(data = systems_spatial %>% filter(county_served %in% east_tx$county_served), 
              lat = ~centroid_lat,
              lng = ~centroid_lon,
              fillColor = ~systempalhb(healthbased_violations_10yr), 
              color = "black", weight = 1.5, label = ~paste(pwsid, healthbased_violations_10yr),
              fillOpacity = 1,
              radius = ~sqrt(estimate_total_pop)/50 + 3)%>%
  addProviderTiles(provider = providers$CartoDB.Voyager)

  systempalpw <- colorNumeric(
  palette = "Blues",
  domain = systems_spatial %>% filter(county_served %in% east_tx$county_served) %>% pull(paperwork_violations_10yr))
  
  leaflet()%>%
  addCircleMarkers(data = systems_spatial %>% filter(county_served %in% east_tx$county_served), 
              lat = ~centroid_lat,
              lng = ~centroid_lon,
              fillColor = ~systempalpw(paperwork_violations_10yr), 
              color = "black", weight = 1.5, label = ~paste(pwsid, paperwork_violations_10yr),
              fillOpacity = 1,
              radius = ~sqrt(estimate_total_pop)/50 + 3)%>%
  addProviderTiles(provider = providers$CartoDB.Voyager)

```
