==> code/4_TX_dw_analysis_socioeconomic.Rmd <==
---
title: "TX Drinking Water Data Analysis - Socioeconomic"
author: "EmmaLi Tsai"
date: "2024-02-23"
last_updated: "2024-04-03"
output: html_document
---
## Loading packages 
```{r}
library(tidyverse)
library(leaflet)
library(sf)
library(aws.s3)
library(corrplot)
```

## Loading data lists 
```{r}
demo <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_demographic_list.RData")

enviro_sw <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_enviro_surface_list.RData")

enviro_gw <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_enviro_ground_list.RData")

wds <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_water_delivery_list.RData")

fin <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_financial_list.RData")

keys <- aws.s3::s3read_using(readRDS, 
                            object = "s3://tech-team-data/state-drinking-water/TX/clean/TX_merging_keys_list.RData")
```

## East TX flag: 
```{r}
# if running stats for all of TX - run this 
census <- demo$census
# for water affordability: 
aff <- fin$TX_affordability_DAC %>%
  as.data.frame() %>%
  select(-c(geometry, estimate_mhi))


# if running stats for just East TX - run this 
census <- demo$census %>%
  filter(east_tx_flag == "yes")
# for water affordability: 
aff <- fin$TX_affordability_DAC %>%
  as.data.frame() %>%
  select(-c(geometry, estimate_mhi)) %>%
  filter(pwsid %in% census$pwsid)
```

## What proportion of rate payers are delinquent in payments?
```{r}
# I don't think we have this data - perhaps we could use the 
# water affordability data as a way to get at this question?
```

## How have water utilities been funded historically? 
NOTE: question has been sunsetted in favor of Policy Link analysis 
```{r}
# metadata for these columns: 
#   - gpr = green project reserve, a capitalization grant that may be used for 
#           projects to address green infrastructure, water, or energy
#           efficiency improvements. This is how much of the project's total 
#           cost is applicable to the GPR. 

# NOTE - pwsid is NOT UNIQUE to entity name! So you may have an entity name 
# (Corix Utilities, for example), that matches to ~8 pwsids

# renaming and fixing pwsids: 
iup <- fin$TX_PPL_SFY14_24 %>%
  rename(pwsid = pws_id)

for(i in 1:nrow(iup)){
  # fill ones with 7 characters, which we can confidently fix: 
  if(nchar(iup$pwsid[i]) %in% c(7)) {
      iup$pwsid[i] <- paste0("TX", iup$pwsid[i], collapse = "")
  }
}

# There's a couple of pwsids in sfy14 that have very ugly pwsids that can't 
# be fixed using the loop above - will require some string matching: 
# creating a flag for PWSIDS that need to be checked 
iup <- iup %>%
  mutate(pwsid_flag = case_when(
    nchar(pwsid) != 9 ~ "flag", 
    pwsid == "TXPending" ~ "flag", 
    TRUE ~ "NA"
  )) %>%
  mutate(pwsid = toupper(pwsid), 
         pwsid = str_squish(pwsid),
         pwsid = gsub(" ", "", pwsid))

# how many do we have to match? 124 
# let's see if we can match with previous name combos: 
all_pwsid_name_combos <- iup %>%
  select(pwsid, entity, population, pwsid_flag) %>%
  filter(pwsid_flag != "flag") %>%
  select(-pwsid_flag) %>%
  unique()

# what are the matching pwsid names from the sabs dataset?
pwsid_names <- demo$census %>%
  as.data.frame() %>%
  select(pwsid, pws_name) %>%
  unique()

iup_sab_names <- merge(all_pwsid_name_combos, pwsid_names, 
                       by = "pwsid", all.x = T)

# merging back with iups: 
iup_tidy <-  merge(iup, iup_sab_names, by = c("entity", "population"), 
                   all.x = T) %>%
  mutate(pwsid_tidy = case_when(
    (nchar(pwsid.x != 9) | (pwsid.x == "TXPENDING")) ~ pwsid.y, 
    TRUE ~ pwsid.x
  ))

# really ~ 50 entities that didn't match nicely 
iup_tidy %>%
  filter(is.na(pwsid.y)) %>%
  count(entity)


### Grabbing basic stats: ###
# Did they apply (count)? 
# Were they awarded (count)
# How much they were awarded (amount)

main_stats <- iup_tidy %>%
  # fixing two TBD values from TXPENDING
  mutate(project_cost = gsub("TBD*", "", project_cost), 
         gpr = gsub("TBD*", "", gpr), 
         project_cost = gsub("\\*", "", project_cost), 
         gpr = gsub("\\*", "", gpr), 
         project_cost_tidy = readr::parse_number(project_cost), 
         gpr_tidy = readr::parse_number(gpr), 
         disadv = readr::parse_number(disadv), 
         disadv = case_when(is.na(disadv) ~ 0, 
                                  TRUE ~ disadv)) %>%
  group_by(sfy, invited) %>%
  summarize(total_applied = length(unique(pwsid_tidy)), 
            amt_awarded = mean(project_cost_tidy, na.rm = T), 
            gpr_awarded = mean(gpr_tidy, na.rm = T), 
            mean_dac = mean(disadv)) %>%
  mutate(invited = case_when( 
    invited == "Yes" ~ "Invited", 
    TRUE ~ "Not Invited"
    )) 
  # pivot_wider(names_from = invited, values_from = total_applied:gpr_awarded) %>%
  # select(-c("amt_awarded_Not Invited", "gpr_awarded_Not Invited")) %>%
  # mutate(total_applied = `total_applied_Invited` + `total_applied_Not Invited`)

main_stats_long <- pivot_longer(main_stats, total_applied:mean_dac)
main_stats_long$name <- as.factor(main_stats_long$name)
levels(main_stats_long$name) <- c("Project Cost ($)", 
                                   "GPR ($)", 
                                   "Mean %DAC", 
                                   "Number of Projects")

ggplot(main_stats_long, aes(x = sfy, y = value, group = invited, 
                            color = invited)) + 
  geom_point() + 
  facet_wrap(~name, scales = "free") + 
  geom_line() + 
  scale_color_discrete(name = "") + 
  theme_minimal() + 
  theme(axis.text.x = element_text(angle = 90,
                                   vjust = 0.5,
                                   hjust=1)) + 
  labs(x = "State Fiscal Year", 
       y = "Value")


# doing the same for East TX: 
etx_pwsids <- keys$analysis_keys %>%
  filter(east_tx_flag == "yes")
etx_main_stats <- iup_tidy %>%
  filter(pwsid_tidy %in% etx_pwsids$pwsid) %>%
  # fixing two TBD values from TXPENDING
  mutate(project_cost = gsub("TBD*", "", project_cost), 
         gpr = gsub("TBD*", "", gpr), 
         project_cost = gsub("\\*", "", project_cost), 
         gpr = gsub("\\*", "", gpr), 
         project_cost_tidy = readr::parse_number(project_cost), 
         gpr_tidy = readr::parse_number(gpr), 
         disadv = readr::parse_number(disadv), 
         disadv = case_when(is.na(disadv) ~ 0, 
                                  TRUE ~ disadv)) %>%
  group_by(sfy, invited) %>%
  summarize(total_applied = length(unique(pwsid_tidy)), 
            amt_awarded = mean(project_cost_tidy, na.rm = T), 
            gpr_awarded = mean(gpr_tidy, na.rm = T), 
            mean_dac = mean(disadv)) %>%
  mutate(invited = case_when( 
    invited == "Yes" ~ "Invited", 
    TRUE ~ "Not Invited"
    )) 
  # pivot_wider(names_from = invited, values_from = total_applied:gpr_awarded) %>%
  # select(-c("amt_awarded_Not Invited", "gpr_awarded_Not Invited")) %>%
  # mutate(total_applied = `total_applied_Invited` + `total_applied_Not Invited`)

etx_main_stats_long <- pivot_longer(etx_main_stats, total_applied:mean_dac)
etx_main_stats_long$name <- as.factor(etx_main_stats_long$name)
levels(etx_main_stats_long$name) <- c("Project Cost ($)", 
                                      "GPR ($)",  
                                      "Mean %DAC", 
                                      "Number of Projects")

ggplot(etx_main_stats_long, aes(x = sfy, y = value, group = invited, 
                            color = invited)) + 
  geom_point() + 
  facet_wrap(~name, scales = "free") + 
  geom_line() + 
  scale_color_discrete(name = "") + 
  theme_minimal() + 
  theme(axis.text.x = element_text(angle = 90,
                                   vjust = 0.5,
                                   hjust=1)) + 
  labs(x = "State Fiscal Year", 
       y = "Value")
```

## What races/ethnicities are served by what system? 
```{r}
# grabbing race, ethnicity, and MHI stats from the census: 
race_stats <- census %>%
  select(pwsid, estimate_white_alone_per:estimate_mixed_alone_per, 
         estimate_poc_alone_per, estimate_hisp_alone_per, estimate_mhi) %>%
  as.data.frame(.) %>%
  select(-geometry) 

# grabbing summary stats for communities served by water systems: 
colMeans(race_stats %>% select(-pwsid), na.rm = TRUE)
summary(census)

# formatting for plotting: 
race_long <- race_stats %>%
  pivot_longer(., estimate_white_alone_per:estimate_mhi)
race_long$name <- as.factor(race_long$name)
levels(race_long$name) <- c("AIAN", "Asian", "Black", 
                            "Latino/a", "MHI", "Mixed", 
                            "NAPI", "Other", "POC", 
                            "White")

# histograms of % each race, ethnicity, and MHI bin: 
options(scipen=10000)
ggplot(race_long, aes(x = value, color = name, fill = name)) + 
  geom_histogram() + 
  facet_wrap(~name, scales = "free", ncol = 5) + 
  theme_minimal() +
  scale_color_discrete(name = "") + 
  scale_fill_discrete(name = "") + 
  theme(legend.position = "bottom") + 
  labs(x = "Percent", y = "Count of PWSIDs") + 
  theme(axis.text.x = 
          element_text(angle = 90, 
                       vjust = 0.5, 
                       hjust=1)) 

# investigating relationships between race, ethnicity, and MHI with 
# health-based violations: 
analysis_keys <- keys$analysis_keys
race_viols_wide <- merge(race_stats, analysis_keys, by = "pwsid") 
race_viols <- merge(race_long, analysis_keys, by = "pwsid") 

race_viol_long <- pivot_longer(race_viols, cols = paperwork_violations_10yr:total_violations_5yr, 
                               names_to = "viol_type", values_to = "total_viols")

ggplot(race_viol_long %>% filter(viol_type %in% c("healthbased_violations_10yr", 
                                                  "healthbased_violations_5yr")), 
       aes(x = value, y = total_viols, color = viol_type)) +
  geom_point() + 
  facet_wrap(~name, scales = "free") + 
  scale_color_discrete(name = "") +
  labs(x = "% Population of Race", y = "Total Violations") + 
  theme_minimal() + 
  theme(legend.position = "bottom")

# calculating and distribution of violations across different race bins: 
races <- race_viols_wide %>%
  select(estimate_white_alone_per:estimate_hisp_alone_per) %>%
  names()
race_viol_breakdown <- data.frame()
for(i in 1:length(races)){
  race_i <- races[i]
  message(race_i)
  # investigating how the data look across quantiles: 
  race_alone_bins <- race_viols_wide %>%
    as.data.frame() %>%
    select(-geometry) %>%
    mutate(poc_bin = cut(!!sym(race_i), 
                         breaks = c(unique(quantile(!!sym(race_i), 
                                             probs = seq(0, 1, by = 0.20),
                                             na.rm = T))), 
                         include.lowest = TRUE)) %>%
    group_by(poc_bin) %>%
    summarize(total_utilities = n(),
              paperwork_10yr = mean(paperwork_violations_10yr), 
              paperwork_5yr = mean(paperwork_violations_5yr), 
              health_10yr = mean(healthbased_violations_10yr), 
              health_5yr = mean(healthbased_violations_5yr), 
              mean_mhi = mean(estimate_mhi.x, na.rm = T)) %>%
    mutate(race = race_i, 
           quantile = ntile(poc_bin, n = length(unique(poc_bin)))) 
  race_viol_breakdown <- rbind(race_viol_breakdown, race_alone_bins)
}
race_viol_breakdown

# taking the df above and just focusing on health-based violations: 
race_5yr_healthbased <- race_viol_breakdown %>%
  select(poc_bin, health_5yr, race) %>%
  pivot_wider(., names_from = race, values_from = health_5yr) 
names(race_5yr_healthbased) <- c("%Race", "White_Viol", "Black_Viol", 
                                 "AIAN_Viol", "Asian_Viol", "NAPI_Viol", 
                                 "Other_Viol", "Mixed_Viol", "POC_Viol", 
                                 "Hisp_Viol")
# this table is rather ugly but is a helpful visual aid for investigating 
# violations by quantile: 
race_5yr_healthbased

# plotting %POC: 
census_sf <- census %>%
  st_transform('+proj=longlat +datum=WGS84') %>%
  filter(east_tx_flag == "yes")

poc_pal <- colorNumeric(
  palette = viridis::viridis(20),
  domain = census_sf$estimate_poc_alone_per)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = census_sf,
              fillOpacity = 0.9,
              # opacity = 0.9,
              # stroke = TRUE,
              # color = "black",
              color = ~poc_pal(estimate_poc_alone_per),
              # fillColor = ~poc_pal(estimate_poc_alone_per),
              weight = 1,
              label = paste0("pwsid: ", census_sf$pwsid,
                             "; poc: ", census_sf$estimate_poc_alone_per)) %>%
  addLegend("bottomright",
            pal = poc_pal,
            values =  census_sf$estimate_poc_alone_per,
            title = "% POC",
            opacity = 1)

## comparison with East TX: ####################################################
# making another histogram showing racial breakdown histograms by comparing 
# east TX: 
race_stats_TX <- demo$census %>%
  select(pwsid, 
         estimate_white_alone_per:estimate_mixed_alone_per, 
         estimate_poc_alone_per, estimate_hisp_alone_per, estimate_mhi) %>%
  as.data.frame(.) %>%
  select(-geometry) %>%
  pivot_longer(., estimate_white_alone_per:estimate_mhi) %>%
  mutate(flag = "TX")

race_stats_eTX <- demo$census %>%
  select(pwsid, east_tx_flag, estimate_white_alone_per:estimate_mixed_alone_per, 
         estimate_poc_alone_per, estimate_hisp_alone_per, estimate_mhi) %>%
  as.data.frame(.) %>%
  select(-geometry) %>%
  filter(east_tx_flag == "yes") %>%
  select(-east_tx_flag) %>%
  pivot_longer(., estimate_white_alone_per:estimate_mhi) %>%
  mutate(flag = "East TX")

all_race <- rbind(race_stats_TX, race_stats_eTX)
all_race$name <- as.factor(all_race$name)
levels(all_race$name) <- c("AIAN", "Asian", "Black", 
                           "Latino/a", "MHI", "Mixed", 
                           "NAPI", "Other", "POC", 
                           "White")

ggplot(all_race, aes(x = value, color = flag, fill = flag)) + 
  geom_histogram(data = subset(all_race, flag == "TX"), 
                 postion = "identity", 
                 alpha = 0.1) + 
  geom_histogram(data = subset(all_race, flag == "East TX"), 
                 postion = "identity", 
                 alpha = 0.1) + 
  facet_wrap(~name, scales = "free", ncol = 5) + 
  theme_minimal() +
  scale_color_discrete(name = "") + 
  scale_fill_discrete(name = "") + 
  theme(legend.position = "bottom") + 
  labs(x = "Percent", y = "Count of PWSIDs") + 
  theme(axis.text.x = 
          element_text(angle = 90, 
                       vjust = 0.5, 
                       hjust=1)) 

# Overall summary: ############################################################
# - Texas is pretty segregated - % white alone is generally greater than 75%, 
#   and very little mixing between white and other races (i.e., the spread of 
#   races looks parabolic)

# - White alone: most fall into the 70-90% white category. Mean health-based 
#   viols over the past 10 years increase as % white increases, but this is also 
#   because there are more utilities that fall within this category. 
#   Same trend is also found with 5-yr health-based violations

# - Black alone: most utilities fall within the 0-10% black. Mean health-based
#   violations over the past 10 years are highest from 0-20% black communities
#   but also peaks around 30-40% and 70-80% (but this n = 1). 

# - Asian alone: similar to findings above, average number of health-based 
#   violations over the past 10 years are highest around 0-10% Asian, same 
#   trend is found with 5-yr violations 

# - Latino/Latina: 30-40% latino/latina communities have fairly high rate of 
#   mean violations, but also larger peak around 0-20%, and 90-100%

# - MHI - generally centered around ~60k - and also has a lot of health-based 
#   violations that mirror this distribution 
# 
# When looking at quanitles: 
#   - Utilities that fall within the highest 20% of percent white (89.2- 100%) 
#     have more mean health-based violations (1.72) 
# 
#   - Utilities that fall within 60-80% quantile of latino/a populations have 
#     the highest mean health-based violations. 

```


## What are the education levels of those served? 
```{r}
# grabbing some basic stats of various education categories: 
edu_cats <- census %>%
  select(pwsid, estimate_no_school_per, estimate_high_school_per, 
         estimate_bachelors_per, estimate_prof_degree_per)
summary(edu_cats)

# grabbing % of population in each category: 
edu_level <- edu_cats %>%
  as.data.frame() %>%
  select(-geometry)

# merging with analysis keys for main stats of interest: 
analysis_keys <- keys$analysis_keys
edu_level_analysis <- merge(edu_level, analysis_keys, by = "pwsid") 

# what do these distributions look like?
edu_level_analysis_long <- edu_level_analysis %>%
  select(pwsid, estimate_no_school_per:estimate_prof_degree_per) %>%
  pivot_longer(estimate_no_school_per:estimate_prof_degree_per)
ggplot(edu_level_analysis_long, aes(x = value, fill = name)) + 
  geom_histogram() + 
  theme_minimal()

# grabbing a table to see how these relate to education levels: 

# calculating and distribution of violations across different race bins: 
edu_categories <- edu_level_analysis %>%
  select(estimate_no_school_per:estimate_prof_degree_per) %>%
  names()
edu_viol_breakdown <- data.frame()
for(i in 1:length(edu_categories)){
  edu_i <- edu_categories[i]
  message(edu_i)
  # investigating how the data look across quantiles: 
  edu_bins <- edu_level_analysis %>%
    as.data.frame() %>%
    select(-geometry) %>%
    mutate(edu_bin = cut(!!sym(edu_i), 
                         breaks = c(unique(quantile(!!sym(edu_i), 
                                             probs = seq(0, 1, by = 0.20),
                                             na.rm = T))), 
                         include.lowest = TRUE)) %>%
    group_by(edu_bin) %>%
    summarize(total_utilities = n(),
              paperwork_10yr = mean(paperwork_violations_10yr), 
              paperwork_5yr = mean(paperwork_violations_5yr), 
              health_10yr = mean(healthbased_violations_10yr), 
              health_5yr = mean(healthbased_violations_5yr), 
              mean_mhi = mean(estimate_mhi, na.rm = T)) %>%
    mutate(race = race_i, 
           quantile = ntile(edu_bin, n = length(unique(edu_bin)))) 
  edu_viol_breakdown <- rbind(edu_viol_breakdown, edu_bins)
}
edu_viol_breakdown


# basic stats: 
edu_level_analysis %>%
  filter(estimate_no_school_per >= 5) %>%
  count()
edu_level_analysis %>%
  filter(estimate_prof_degree_per >= 5) %>%
  count()

# what's the correlation between schooling and mhi?
mhi_school <- edu_level_analysis %>%
  pivot_longer(estimate_no_school_per:estimate_prof_degree_per)
mhi_school$name <- as.factor(mhi_school$name)
levels(mhi_school$name) <- c("Bachelors",
                             "High School", 
                             "No Schooling", 
                             "Professional Degree")
ggplot(mhi_school, aes(x = value, y = estimate_mhi)) + 
  geom_point() + 
  facet_wrap(~name, scales = "free") + 
  geom_smooth(color = "grey") + 
  geom_point(data = mhi_school %>% filter(healthbased_violations_5yr > 0), 
             color = "red") + 
  labs(x = "% Education Category", y = "Estimate MHI") + 
  theme_minimal()
# yay! the data make sense! 

## comparison with East TX: ####################################################
edu_level_TX <- demo$census %>%
  select(pwsid, estimate_no_school_per:estimate_prof_degree_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(flag = "TX")

edu_level_eTX <- demo$census %>%
  select(pwsid, east_tx_flag, estimate_no_school_per:estimate_prof_degree_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  filter(east_tx_flag == "yes") %>%
  select(-east_tx_flag) %>%
  mutate(flag = "East TX")

all_edu <- rbind(edu_level_TX, edu_level_eTX)
all_edu <- pivot_longer(all_edu, estimate_no_school_per:estimate_prof_degree_per)
all_edu$name <- as.factor(all_edu$name)
levels(all_edu$name) <- c("Bachelors",
                          "High School", 
                          "No Schooling", 
                          "Professional Degree")

ggplot(all_edu, aes(x = name, y = value, 
                    fill = flag)) + 
  geom_boxplot(alpha = 0.5) + 
  scale_fill_discrete(name = "") +
  theme_minimal()  + 
  labs(x = "", y = "Percent of Population")

# Overall summary: ############################################################
# - As the % of people without schooling increases, MHI decreases
# - As the % of people with professional degrees increases, MHI increases 
#   but appears to level out after the percentage of people with prof degrees 
#   reaches ~10% 
# - 72 (1.6%) pwsids serve communities where >= 5% of the community has had no form of 
#   schooling. 
# - 54 (1.2%) pwsids serve communities where >= 5% of the community has professional 
#   degrees, and visually it appears that communities that are more educated
#   have less health-based violations 
# - Most pwsids fall within the 0-2% for both of these categories. Communities 
#   with larger % no schooling population (0.01-4)% are served by utilities with 
#   ~1 more mean health-based violation than communities where 0.01-4% have 
#   professional degrees. Interestingly, the largest mean health-based violations 
#   occurred in communities where 8-10% had professional degrees, but also only 
#   15 pwsids fell within this category. A similar rise in health-based 
#   violations was also discovered in communities with a similar % of 
#   individuals without schooling. 

#     By quantiles: 
#       - Utilities that fall within the lowest 20% quantile of professional 
#         degrees (0 - 0.05%) have the most mean health-based and paperwork 
#         violations over the past 5 years. As quantile increases in 
#         professional degrees, MHI increases. 
# 
#       - Utilities that fall within the 40-60% quantile of populations 
#         without schooling (0.65 - 1.12%) have the most mean health-based 
#         violations. As quantile increase, MHI decreases and mean paperwork 
#         violations over the past 5 and 10 years increase.  

```

## What languages are spoken by those served?
```{r}
# grabbing language spoken at home and merging with key stats of interest: 
lang <- census %>%
  select(pwsid, estimate_only_english_per:estimate_other_lang_per) %>%
  as.data.frame() %>%
  select(-geometry)
analysis_keys <- keys$analysis_keys
lang_analysis <- merge(lang, analysis_keys, by = "pwsid") 

# plot to see what the data look like: 
lang_analysis_long <- pivot_longer(lang_analysis, 
                                   cols = c("estimate_only_english_per", 
                                            "estimate_other_lang_per"))
ggplot(lang_analysis,
       aes(x = estimate_only_english_per)) +
  geom_histogram()

# plots to investigate relationships between language spoken at home and 
# other metrics - basically confirming stuff we already know: 
# how does this compare to % POC?
ggplot(lang_analysis, aes(x = estimate_only_english_per, 
                          y = estimate_poc_alone_per)) + 
  geom_point() + 
  geom_smooth(color = "grey") + 
  theme_minimal() +
  geom_point(data = lang_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  geom_smooth(data = lang_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") + 
  labs(x = "% Only Speak English", y = "% POC") 

# language and mhi
ggplot(lang_analysis, aes(x = estimate_only_english_per, 
                          y = estimate_mhi)) + 
  geom_point() + 
  geom_smooth(color = "grey") + 
  theme_minimal() +
  geom_point(data = lang_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  geom_smooth(data = lang_analysis %>% filter(healthbased_violations_5yr>0), 
              color = "red") + 
  labs(x = "% Only Speak English", y = "MHI") 

# checking out binned results in table form: 
lang_analysis %>%
    as.data.frame() %>%
    select(-geometry) %>%
    mutate(lang_bin = cut(estimate_only_english_per, seq(0, 100, by = 10))) %>%
    group_by(lang_bin) %>%
    summarize(total_utilities = n(),
              paperwork_10yr = mean(paperwork_violations_10yr), 
              paperwork_5yr = mean(paperwork_violations_5yr), 
              health_10yr = mean(healthbased_violations_10yr), 
              health_5yr = mean(healthbased_violations_5yr)) 

# by quantile: 
lang_analysis %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(lang_bin = cut(estimate_only_english_per, 
                       breaks = c(quantile(estimate_only_english_per, 
                                           probs = seq(0, 1, by = 0.20), 
                                           na.rm = TRUE)), 
                       include.lowest = TRUE)) %>%
  group_by(lang_bin) %>%
  summarize(total_utilities = n(),
            paperwork_10yr = mean(paperwork_violations_10yr), 
            paperwork_5yr = mean(paperwork_violations_5yr), 
            health_10yr = mean(healthbased_violations_10yr), 
            health_5yr = mean(healthbased_violations_5yr), 
            mean_mhi = mean(estimate_mhi, na.rm = T)) %>%
  mutate(quantile = c("0-20","20-40","40-60","60-80","80-100", "NA")) %>%
  relocate(quantile, .after = lang_bin)

# how many serve populations >= 60% speak only english at home?
lang_analysis %>%
  filter(estimate_only_english_per >= 60) %>%
  count()

# comparisons with East TX: ###################################################
# can we do boxplots to show comparisons with east TX:
english_TX <- demo$census %>%
  select(pwsid, estimate_only_english_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(flag = "TX")

english_eTX <- demo$census %>%
  select(pwsid, east_tx_flag, estimate_only_english_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  filter(east_tx_flag == "yes") %>%
  select(-east_tx_flag) %>%
  mutate(flag = "East TX")

all_lang_long <- rbind(english_TX, english_eTX)

ggplot(all_lang_long, aes(x = flag, y = estimate_only_english_per,
                    fill = flag)) + 
  geom_boxplot(alpha = 0.5) + 
  scale_fill_discrete(name = "") +
  theme_minimal()  + 
  labs(y = "Percent Speaks Only English at Home", x = "") + 
  theme(legend.position = "none")


# Overall findings: ########################################################### 
#   - to no surprise, the higher % that speak English at home, the lower the 
#     % POC. Based on the trends, it also appears that systems with health-based  
#     violations are concentrated around systems that serve communities where 
#     >= 75% of the community speaks only English. 

#   - as the percentage of people that speak only English at home increases, the 
#     MHI increases 

#   - 78.8% of sabs serve communities where >= 60% of people speak only English 

#   - When a utility serves populations where 0-10% of the community speaks 
#     only english, the average number of health-based violations is highest 
#     by about 1. Another peak is found around communities where 40-50% speak 
#     only English, and generally any communties that serve communities where 
#     more than 60% of the community speaks only English, but that might again 
#     be an artifact since there are more utilities that fall into this 
#     category. 
# 
#   - By quantile:
#       - Utilities that fall within the higher quantiles (40-100) of 
#         the percent of population that speaks only english generally have 
#         more mean heath-based violations over the last 5 years and lower MHI
```

## What nationality/immigration status reflect those seved? 
```{r}
# grabbing immigration status: 
immi_status <- census %>%
  select(pwsid, estimate_foreign_per) %>%
  as.data.frame() %>%
  select(-geometry)

# checking out what the distribution looks like: 
ggplot(immi_status, aes(x = estimate_foreign_per)) + 
  geom_histogram()

# how does immigration status relate to poc, mhi, and sdwis violations: 
analysis_keys <- keys$analysis_keys
immi_analysis <- merge(immi_status, analysis_keys, by = "pwsid") 
ggplot(immi_analysis, aes(x = estimate_foreign_per, 
                          y = estimate_poc_alone_per)) + 
  geom_point() + 
  geom_smooth() + 
  geom_smooth(color = "grey") + 
  theme_minimal() +
  geom_point(data = immi_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  geom_smooth(data = immi_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") + 
  labs(x = "% Immigrant Status", y = "%POC") 

# how does immigrant status relate to mhi?
ggplot(immi_analysis, aes(x = estimate_foreign_per, 
                          y = estimate_mhi)) + 
  geom_point() + 
  geom_smooth() + 
  geom_smooth(color = "grey") + 
  theme_minimal() +
  geom_point(data = immi_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  geom_smooth(data = immi_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  labs(x = "% Immigrant Status", y = "MHI")

# checking out binned results in table form: 
immi_analysis %>%
    as.data.frame() %>%
    select(-geometry) %>%
    mutate(immi_bin = cut(estimate_foreign_per, seq(0, 100, by = 10))) %>%
    group_by(immi_bin) %>%
    summarize(total_utilities = n(),
              paperwork_10yr = mean(paperwork_violations_10yr), 
              paperwork_5yr = mean(paperwork_violations_5yr), 
              health_10yr = mean(healthbased_violations_10yr), 
              health_5yr = mean(healthbased_violations_5yr)) 

# by quantile: 
immi_analysis %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(immi_bin = cut(estimate_foreign_per, 
                       breaks = c(quantile(estimate_foreign_per, 
                                           probs = seq(0, 1, by = 0.20), 
                                           na.rm = TRUE)), 
                       include.lowest = TRUE)) %>%
  group_by(immi_bin) %>%
  summarize(total_utilities = n(),
            paperwork_10yr = mean(paperwork_violations_10yr), 
            paperwork_5yr = mean(paperwork_violations_5yr), 
            health_10yr = mean(healthbased_violations_10yr), 
            health_5yr = mean(healthbased_violations_5yr), 
            mean_mhi = mean(estimate_mhi, na.rm = T)) %>%
  mutate(quantile = c("0-20","20-40","40-60","60-80","80-100", "NA")) %>%
  relocate(quantile, .after = immi_bin)


# comparisons with East TX: ###################################################
immi_TX <- demo$census %>%
  select(pwsid, estimate_foreign_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(flag = "TX")

immi_analysis_etx <- demo$census %>%
  select(pwsid, east_tx_flag, estimate_foreign_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  filter(east_tx_flag == "yes") %>%
  mutate(flag = "East TX") %>%
  select(-east_tx_flag)

all_immi <- rbind(immi_TX, immi_analysis_etx)

# merging with analysis keys for main stats of interest: 
analysis_keys <- keys$analysis_keys
all_immi <- merge(all_immi, analysis_keys, by = "pwsid") %>%
  select(pwsid, estimate_foreign_per, flag)

ggplot(all_immi, aes(x = flag, y = estimate_foreign_per,
                    fill = flag)) + 
  geom_boxplot(alpha = 0.5) + 
  scale_fill_discrete(name = "") +
  theme_minimal()  + 
  labs(y = "% Immigrant Status", x = "") + 
  theme(legend.position = "none")

# some basic stats: 
mean(all_immi$estimate_foreign_per, na.rm = T)
median(all_immi$estimate_foreign_per, na.rm = T)

etx_immi <- all_immi %>% filter(flag == "East TX")
mean(etx_immi$estimate_foreign_per, na.rm = T)
median(etx_immi$estimate_foreign_per, na.rm = T)


# Overall findings: ############################################################
# - 83.14% of pwsids serve communities where less than 20% immigrated from a 
#   different county. 

# - to no surprise, pwsids with larger % immigrant status also have 
#   higher % POC - this is a pretty strong relationship. 

# - pwsids with larger % immigrant populations generally have smaller MHI, 
#   but this correlation is pretty weak. When looking pwsids with health-
#   based violations over the past 5 years, they tend to have lower MHI
#   and smaller % population that are immigrants. 

# - pwsids with <= 20% of the community of immigrant status generally have 
#   more health-based violations over the past 5 (and particularly 10) years, 
#   but that might be due to the fact there are more pwsids that fall into 
#   this category. Paperwork violations also tend to be higher with utilities 
#   that serve communities where small % of the community is from out of the 
#   county. 

# By quantile: 
# - Utilities that fall within the lowest quantile of % immigrant population 
#   have more mean health-based and paperwork violations over the past 5 and 10 
#   years. 
# - Utilities that fall within the highest quantile of % immigrant 
#   population have larger MHI
```

## How affordable is the water? 
```{r}
# NOTE: based on past IUPs, here is how hcf relates to principal forgiveness as
# a % of DWSRF funded project costs remaining after subtracting other 
# DWSRF princiapl forgiveness: 
# https://www.twdb.texas.gov/financial/programs/DWSRF/doc/SFY2020/SFY2020_DWSRF_IUP.pdf
#   - hcf difference
#       - >=0 and < 1.5% --> 30% 
#       - >= 1.5 and < 3 % --> 50% 
#       - >= 3% --> 70%

# NOTE: data here are also conservative - hcf is further modified by 
# unemployment rate hcf and population decline hcf, which can only 
# increase the chance of being identified as disadvantaged 
# NOTE: the water affordability data is NOT comprehensive - only has ~706 
# unique pwsids 

# how does hcf  relate to poc, mhi, and sdwis violations: 
analysis_keys <- keys$analysis_keys
aff_analysis <- merge(aff, analysis_keys, by = "pwsid") 
mean(aff_analysis$total_water_year) 
# Mean just total water: 
# East TX: $511.2519
# TX: $545.5747


# how does household cost factor relate to %POC?
ggplot(aff_analysis, aes(x = hcf*100, 
                          y = estimate_poc_alone_per)) + 
  geom_point() + 
  geom_smooth() + 
  geom_smooth(color = "grey") + 
  theme_minimal() +
  geom_point(data = aff_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  geom_smooth(data = aff_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") + 
  labs(x = "Household Cost Factor", y = "%POC") 

# how does household cost factor relate to MHI? should be a strong relationship
ggplot(aff_analysis, aes(x = hcf*100,
                          y = estimate_mhi)) +
  geom_point() +
  geom_smooth() +
  geom_smooth(color = "grey") +
  theme_minimal() +
  geom_point(data = aff_analysis %>% filter(healthbased_violations_5yr>0),
             color = "red") +
  geom_smooth(data = aff_analysis %>% filter(healthbased_violations_5yr>0),
             color = "red")  + 
  labs(x = "Household Cost Factor", y = "MHI")

# mean and median: 
mean(aff_analysis$hcf*100, na.rm = T)
median(aff_analysis$hcf*100, na.rm = T)

# checking out binned results in table form: 
aff_analysis %>%
    as.data.frame() %>%
    select(-geometry) %>%
    mutate(aff_bin = cut(hcf, c(0, 0.01, 0.02, 0.03, 0.04, 0.05))) %>%
    group_by(aff_bin) %>%
    summarize(total_utilities = n(),
              paperwork_10yr = mean(paperwork_violations_10yr), 
              paperwork_5yr = mean(paperwork_violations_5yr), 
              health_10yr = mean(healthbased_violations_10yr), 
              health_5yr = mean(healthbased_violations_5yr)) 

# by quantile: 
aff_analysis %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(aff_bin = cut(hcf, 
                       breaks = c(quantile(hcf, 
                                           probs = seq(0, 1, by = 0.25), 
                                           na.rm = TRUE)), 
                       include.lowest = TRUE)) %>%
  group_by(aff_bin) %>%
  summarize(total_utilities = n(),
            paperwork_10yr = mean(paperwork_violations_10yr), 
            paperwork_5yr = mean(paperwork_violations_5yr), 
            health_10yr = mean(healthbased_violations_10yr), 
            health_5yr = mean(healthbased_violations_5yr), 
            mean_mhi = mean(estimate_mhi, na.rm = T), 
            mean_poc = mean(estimate_poc_alone_per, na.rm = T)) 


# comparisons with East TX #####################################################
# creating data frames for TX - can filter for east TX from here:  
aff_TX <- fin$TX_affordability_DAC %>%
  as.data.frame() %>%
  as.data.frame() %>%
  select(-c(geometry, estimate_mhi))
aff_analysis <- merge(aff_TX, analysis_keys, by = "pwsid") 

# what does hcf look like spatially? 
aff_analysis_sf <- aff_analysis %>%
  st_as_sf() %>%
  st_transform('+proj=longlat +datum=WGS84') %>%
  filter(east_tx_flag == "yes")

# mapping household cost factor across TX: 
hcf_pal <- colorNumeric(
  palette = c("#003049","#6b2c39", "#d62828", "#f77f00", "#fcbf49"),
  domain = aff_analysis_sf$hcf*100)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = aff_analysis_sf,  
              # %>% filter(east_tx_flag == "yes"),
              fillOpacity = 0.9,
              # opacity = 9,
              color = ~hcf_pal(hcf*100),
              weight = 1,
              label = paste0("pwsid: ", aff_analysis_sf$pwsid,
                "; hcf: ", round(aff_analysis_sf$hcf, 2))) %>%
  addLegend("bottomright",
            pal = hcf_pal,
            title = "HCF",
            values = aff_analysis_sf$hcf*100,
            opacity = 1)

# what do water rates look like spatially?
aff_analysis_sf <- aff_analysis_sf %>%
  mutate(total_water = total_water_year + total_sewer_year)
water_rate_pal <- colorNumeric(
  palette = c("#003049","#6b2c39", "#d62828", "#f77f00", "#fcbf49"),
  domain = aff_analysis_sf$total_water)

# boundaries of all sabs: 
sabs <- demo$census %>%
  st_transform('+proj=longlat +datum=WGS84') %>%
  filter(east_tx_flag == "yes")

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = sabs, 
              opacity = 9,
              color = "grey",
              weight = 1) %>%
  addPolygons(data = aff_analysis_sf, 
              fillOpacity = 0.9,
              # opacity = 9,
              color = ~water_rate_pal(total_water),
              weight = 1,
              label = paste0("pwsid: ", aff_analysis_sf$pwsid,
                "; hcf: ", round(aff_analysis_sf$total_water, 2))) %>%
  addLegend("bottomright",
            pal = water_rate_pal,
            values = aff_analysis_sf$total_water,
            title = "Annual Water Rate ($)",
            opacity = 1)


# grabbing mean stats for water affordability in TX and east TX: 
mean(aff_analysis_sf$total_water_year)
mean(aff_analysis_sf$total_sewer_year)
median(aff_analysis_sf$total_water_year)
median(aff_analysis_sf$total_sewer_year)

etx_aff <- aff_analysis_sf %>%
  filter(east_tx_flag == "yes")

mean(etx_aff$total_water_year)
mean(etx_aff$total_sewer_year)
median(etx_aff$total_water_year)
median(etx_aff$total_sewer_year)


# how does this compare to mhi? creating a map of MHI across east TX:
census_sf <- demo$census %>%
  st_transform('+proj=longlat +datum=WGS84') %>%
  filter(east_tx_flag == "yes")

mhi_pal <- colorNumeric(
  palette = viridis::viridis(9),
  domain = census_sf$estimate_mhi)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = census_sf,
              fillOpacity = 0.9,
              # opacity = 9,
              # stroke = TRUE, 
              # color = "black", 
              color = ~mhi_pal(estimate_mhi),
              # fillColor = ~mhi_pal(estimate_mhi),
              weight = 1,
              label = paste0("pwsid: ", aff_analysis_sf$pwsid,
                             "; hcf: ", round(aff_analysis_sf$total_water_year, 2))) %>%
  addLegend("bottomright",
            pal = mhi_pal,
            values = census_sf$estimate_mhi,
            title = "MHI",
            opacity = 1)


# box plots to show comparisons with East TX: 
east_tx_pwsids <- demo$census %>% 
  filter(east_tx_flag == "yes")
hcf_TX <- fin$TX_affordability_DAC %>%
  select(pwsid, hcf) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(flag = "TX")

hcf_eTX <- aff_analysis %>%
  as.data.frame() %>%
  select(-geometry) %>%
  filter(pwsid %in% east_tx_pwsids$pwsid) %>%
  select(pwsid, hcf) %>%
  mutate(flag = "East TX")

all_hcf <- rbind(hcf_eTX, hcf_TX)

ggplot(all_hcf, aes(x = flag, y = hcf*100,
                    fill = flag)) + 
  geom_boxplot(alpha = 0.5) + 
  scale_fill_discrete(name = "") +
  theme_minimal()  + 
  labs(y = "Household Cost Factor", x = "") + 
  theme(legend.position = "none")

# what does the cost of water look like in comparison to MHI?
ggplot(aff_analysis, aes(x = total_water_year + 
                           total_sewer_year,
                          y = estimate_mhi)) +
  geom_point() +
  geom_smooth() +
  geom_smooth(color = "grey") +
  theme_minimal() +
  geom_point(data = aff_analysis %>% filter(healthbased_violations_5yr>0),
             color = "red") +
  geom_smooth(data = aff_analysis %>% filter(healthbased_violations_5yr>0),
             color = "red")  + 
  labs(x = "Annual Water Cost (Water + Sewer)", y = "MHI")

# what are the characteristic of various bins of income? 
mhi_analysis <- demo$census %>%
  select(pwsid, east_tx_flag, estimate_mhi) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  left_join(keys$analysis_keys)

mhi_analysis %>%
  as.data.frame() %>%
  # filter(east_tx_flag == "yes") %>%
  select(-geometry) %>%
  mutate(mhi_bin = cut(estimate_mhi, 
                       breaks = c(quantile(estimate_mhi, 
                                           probs = seq(0, 1, by = 0.2), 
                                           na.rm = TRUE)), 
                       include.lowest = TRUE)) %>%
  group_by(mhi_bin) %>%
  summarize(total_utilities = n(),
            paperwork_10yr = mean(paperwork_violations_10yr), 
            paperwork_5yr = mean(paperwork_violations_5yr), 
            health_10yr = mean(healthbased_violations_10yr), 
            health_5yr = mean(healthbased_violations_5yr), 
            mean_poc = mean(estimate_poc_alone_per, na.rm = T)) 


################################################################################
# Checking out LQI 
################################################################################
# checking out lowest quintile of income
mhi <- demo$census %>%
  select(pwsid, estimate_mhi) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  left_join(keys$analysis_keys) %>%
  filter(east_tx_flag == "yes")

# by quintile: 
mhi %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(mhi_bin = cut(estimate_mhi, 
                        breaks = c(quantile(estimate_mhi, 
                                            probs = seq(0, 1, by = 0.20), 
                                            na.rm = TRUE)), 
                        include.lowest = TRUE)) %>%
  group_by(mhi_bin) %>%
  summarize(total_utilities = n(),
            paperwork_10yr = mean(paperwork_violations_10yr), 
            paperwork_5yr = mean(paperwork_violations_5yr), 
            health_10yr = mean(healthbased_violations_10yr), 
            health_5yr = mean(healthbased_violations_5yr), 
            mean_poc = mean(estimate_poc_alone_per, na.rm = T)) %>%
  mutate(quintiles = c("0-20","20-40","40-60","60-80","80-100", "NA")) %>%
  relocate(quintiles, .after = mhi_bin)
# lower 20% = $23,600 - $43,700 inclusive  

# checking out characteristics of this group: 
lowest_quintile_mhi <- demo$census %>%
  filter(estimate_mhi <= 43700) %>% 
  filter(east_tx_flag == "yes")
# there are 515, 145 of which are in east tx

# interested in seeing where these are: 
mhi_pal_mini <- colorNumeric(
  palette = viridis::viridis(9),
  domain = mhi$estimate_mhi)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = mhi %>% st_as_sf() %>% st_transform('+proj=longlat +datum=WGS84'),
              fillOpacity = 0.9,
              color = ~mhi_pal_mini(estimate_mhi),
              weight = 1,
              label = paste0("pwsid: ", mhi$pwsid)) %>%
  addPolygons(data = lowest_quintile_mhi%>% st_transform('+proj=longlat +datum=WGS84'),
              fillOpacity = 0.9,
              color = "red",
              weight = 1) %>%
  addLegend("bottomright", 
            colors = c("red"),
            labels = c(""),
            title = "Lowest Quintile",
            opacity = 1) %>%
  addLegend("bottomright",
            pal = mhi_pal_mini,
            values = mhi$estimate_mhi,
            title = "MHI",
            opacity = 1) 

# What are these utilities paying for water? 
aff_analysis %>%
    # filter(pwsid %in% lowest_quintile_mhi$pwsid) %>%
  filter(!(pwsid %in% lowest_quintile_mhi$pwsid)) %>%
  mutate(total_water_sewer_year = total_water_year + 
           total_sewer_year) %>%
  summarize(mean_total_water_sewer = mean(total_water_sewer_year), 
            mean_total_water = mean(total_water_year), 
            meean_total_sewer = mean(total_sewer_year))

# How does HCF relate to LQI? 
lqi_hcf <- aff %>%
  left_join(keys$analysis_keys, "pwsid") %>%
  filter(east_tx_flag == "yes")

ggplot(lqi_hcf, aes(x = estimate_mhi, y = total_water_year + total_sewer_year)) + 
  geom_point(aes(color = estimate_poc_alone_per), size = 5, 
             alpha = 0.8) + 
  # scale_color_continuous(name = "HCF") + 
  scale_color_continuous(name = "% POC") + 
  theme_minimal() + 
  labs(x = "Annual Median Household Income ($)", 
       y = "Annual Total Household Water Cost ($)") + 
  geom_vline(xintercept = 43700, lty = "dashed")
# Makes sense - they should be related since HCF is calculated using MHI


# Overall findings: 
# - utilities that serve communities where the water is less affordable 
#   (hcf > 0.02), have strikingly higher mean health-based violations 
#   over the past 5 and 10 years. Mean values are particularly high 
#   between 0.02-0.03 but that might be due to the outlier that had 
#   ~182 (!!!!!) health-based violations over the past five years. 

# - when plotting the data spatially, large cities generally 
#   have low hcf, but rural communities have higher hcf. 

# By quantiles: 
# Utilities that fall into the highest quantile of hcf (0.019 - 0.0449) have 
# the lowest MHI, generally lower %POC,  and the most health-based violations 
# over the past five and ten years. 
```

## What ages reflect those served?
```{r}
# box plot to investigate each age bracket: 
age_level_TX <- demo$census %>%
  select(pwsid, estimate_ageunder_5_per:estimate_age60_61_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(flag = "TX")

age_level_eTX <- demo$census %>%
  select(pwsid, east_tx_flag, estimate_ageunder_5_per:estimate_age60_61_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  filter(east_tx_flag == "yes") %>%
  select(-east_tx_flag) %>%
  mutate(flag = "East TX")

all_age <- rbind(age_level_TX, age_level_eTX)
all_age <- pivot_longer(all_age, estimate_ageunder_5_per:estimate_age60_61_per)
all_age$name <- as.factor(all_age$name)
levels(all_age$name) <- c("Age 18-24", "Age 25-34", 
                          "Age 35-44", "Age 45-54", 
                          "Age 05-17", "Age 55-59", "Age 60-61", 
                          "Age 0-5")

age_order <- c("Age 0-5", "Age 05-17", "Age 18-24", "Age 25-34", 
               "Age 35-44", "Age 45-54","Age 55-59", "Age 60-61") 

# running boxplot: 
ggplot(all_age, aes(x = factor(name, level = age_order),  
                    y = value, 
                    fill = flag)) + 
  geom_boxplot(alpha = 0.5) + 
  scale_fill_discrete(name = "") +
  theme_minimal()  + 
  labs(x = "", y = "Percent of Population") + 
  coord_flip()

# pulling out some summary stats: 
all_age %>%
  group_by(name, flag) %>%
  summarize(mean_val = mean(value, na.rm = T))


# creating a population chart - but need to grab actual numbers: 
age_level_TX_pop <- demo$census %>%
  select(pwsid, estimate_ageunder_5:estimate_age60_61) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(flag = "TX")

age_level_eTX_pop <- demo$census %>%
  select(pwsid, east_tx_flag, estimate_ageunder_5:estimate_age60_61) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  filter(east_tx_flag == "yes") %>%
  select(-east_tx_flag) %>%
  mutate(flag = "East TX")

all_age_pop <- rbind(age_level_TX_pop, age_level_eTX_pop)
all_age_pop <- pivot_longer(all_age_pop, estimate_ageunder_5:estimate_age60_61)
all_age_pop$name <- as.factor(all_age_pop$name)
levels(all_age_pop$name) <- c("Age 18-24", "Age 25-34", 
                              "Age 35-44", "Age 45-54", 
                              "Age 05-17", "Age 55-59", "Age 60-61", 
                              "Age 0-5")

age_order <- c("Age 0-5", "Age 05-17", "Age 18-24", "Age 25-34", 
               "Age 35-44", "Age 45-54","Age 55-59", "Age 60-61") 

all_age_pop %>% mutate( 
  population = ifelse(flag=="TX", value*(-1), 
                      value*1))%>% 
  ggplot(aes(x = factor(name, level = age_order), 
             y = population, fill = flag)) +  
  geom_bar(stat = "identity") + 
  coord_flip() + 
  theme_minimal() + 
  scale_fill_discrete(name = "") + 
  labs(y = "Population", x = "Age Bin")
```

## What unemployment status reflect those served?
```{r}
# box plot to investigate each age bracket: 
unemp_TX <- demo$census %>%
  select(pwsid, estimate_laborforce_unemployed_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(flag = "TX")

unemp_eTX <- demo$census %>%
  select(pwsid, east_tx_flag, estimate_laborforce_unemployed_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  filter(east_tx_flag == "yes") %>%
  select(-east_tx_flag) %>%
  mutate(flag = "East TX")

all_unep <- rbind(unemp_TX, unemp_eTX)

# running boxplot: 
ggplot(all_unep, aes(x = flag,  
                    y = estimate_laborforce_unemployed_per, 
                    fill = flag)) + 
  geom_boxplot(alpha = 0.5) + 
  scale_fill_discrete(name = "") +
  theme_minimal()  + 
  labs(x = "", y = "Percent of Labor Force Unemployed") 

# what does this look like spatially?
census_sf <- demo$census %>%
  st_transform('+proj=longlat +datum=WGS84') %>%
  filter(east_tx_flag == "yes")

unemp_pal <- colorNumeric(
  palette = viridis::viridis(9),
  domain = census_sf$estimate_laborforce_unemployed_per)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = census_sf,
              fillOpacity = 0.9,
              color = ~unemp_pal(estimate_laborforce_unemployed_per),
              weight = 1,
              label = paste0("pwsid: ", census_sf$pwsid, 
                             "; % unemp: ", 
                             census_sf$estimate_laborforce_unemployed_per)) %>%
  addLegend("bottomright",
            pal = unemp_pal,
            values = census_sf$estimate_laborforce_unemployed_per,
            title = "% Unemployment",
            opacity = 1) 

# by quantile: 
unemp_analysis <- demo$census %>%
  # filter(east_tx_flag == "yes") %>%
  select(pwsid, estimate_laborforce_unemployed_per) %>%
  as.data.frame() %>%
  left_join(keys$analysis_keys)
  
unemp_analysis  %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(unemp_bin = cut(estimate_laborforce_unemployed_per, 
                       breaks = c(quantile(estimate_laborforce_unemployed_per, 
                                           probs = seq(0, 1, by = 0.20), 
                                           na.rm = TRUE)), 
                       include.lowest = TRUE)) %>%
  group_by(unemp_bin) %>%
  summarize(total_utilities = n(),
            paperwork_10yr = mean(paperwork_violations_10yr), 
            paperwork_5yr = mean(paperwork_violations_5yr), 
            health_10yr = mean(healthbased_violations_10yr), 
            health_5yr = mean(healthbased_violations_5yr), 
            mean_mhi = mean(estimate_mhi, na.rm = T)) %>%
  mutate(quantile = c("0-20","20-40","40-60","60-80","80-100", "NA")) %>%
  relocate(quantile, .after = unemp_bin)

ggplot(unemp_analysis, aes(x = estimate_laborforce_unemployed_per, 
                          y = estimate_poc_alone_per)) + 
  geom_point() + 
  geom_smooth() + 
  geom_smooth(color = "grey") + 
  theme_minimal() +
  geom_point(data = unemp_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  geom_smooth(data = unemp_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") + 
  labs(x = "% Unemployed", y = "%POC") 

# how does immigrant status relate to mhi?
ggplot(unemp_analysis, aes(x = estimate_laborforce_unemployed_per, 
                          y = estimate_mhi)) + 
  geom_point() + 
  geom_smooth() + 
  geom_smooth(color = "grey") + 
  theme_minimal() +
  geom_point(data = unemp_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  geom_smooth(data = unemp_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  labs(x = "% Unemployed", y = "MHI")
```

## What economic status reflect households served?
```{r}
# box plot to investigate each age bracket: 
hh_pov_TX <- demo$census %>%
  select(pwsid, estimate_hh_below_pov_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(flag = "TX")

hh_pov_eTX <- demo$census %>%
  select(pwsid, east_tx_flag, estimate_hh_below_pov_per) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  filter(east_tx_flag == "yes") %>%
  select(-east_tx_flag) %>%
  mutate(flag = "East TX")

all_hh_pov <- rbind(hh_pov_TX, hh_pov_eTX)

# running boxplot: 
ggplot(all_hh_pov, aes(x = flag,  
                    y = estimate_hh_below_pov_per, 
                    fill = flag)) + 
  geom_boxplot(alpha = 0.5) + 
  scale_fill_discrete(name = "") +
  theme_minimal()  + 
  labs(x = "", y = "Percent of Households Under the Poverty Line") 

# what does this look like spatially?
census <- demo$census %>%
  filter(east_tx_flag == "yes")
hh_pov_pal <- colorNumeric(
  palette = viridis::viridis(9),
  domain = census$estimate_hh_below_pov_per)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = census,
              fillOpacity = 0.9,
              color = ~hh_pov_pal(estimate_hh_below_pov_per),
              weight = 1,
              label = paste0("pwsid: ", census$pwsid, 
                             "; % hh poverty: ", census$estimate_hh_below_pov_per)) %>%
  addLegend("bottomright",
            pal = hh_pov_pal,
            values = census$estimate_hh_below_pov_per,
            title = "% Households in Poverty",
            opacity = 1) 

# by quantile: 
hh_pov_analysis <- demo$census %>%
  filter(east_tx_flag == "yes") %>%
  select(pwsid, estimate_hh_below_pov_per) %>%
  as.data.frame() %>%
  left_join(keys$analysis_keys)
  
hh_pov_analysis  %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(hh_pov_bin = cut(estimate_hh_below_pov_per, 
                       breaks = c(quantile(estimate_hh_below_pov_per, 
                                           probs = seq(0, 1, by = 0.20), 
                                           na.rm = TRUE)), 
                       include.lowest = TRUE)) %>%
  group_by(hh_pov_bin) %>%
  summarize(total_utilities = n(),
            paperwork_10yr = mean(paperwork_violations_10yr), 
            paperwork_5yr = mean(paperwork_violations_5yr), 
            health_10yr = mean(healthbased_violations_10yr), 
            health_5yr = mean(healthbased_violations_5yr), 
            mean_mhi = mean(estimate_mhi, na.rm = T)) %>%
  mutate(quantile = c("0-20","20-40","40-60","60-80","80-100", "NA")) %>%
  relocate(quantile, .after = hh_pov_bin)

# plotting relationships: 
ggplot(hh_pov_analysis, aes(x = estimate_hh_below_pov_per, 
                          y = estimate_poc_alone_per)) + 
  geom_point() + 
  geom_smooth() + 
  geom_smooth(color = "grey") + 
  theme_minimal() +
  geom_point(data = hh_pov_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  geom_smooth(data = hh_pov_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") + 
  labs(x = "% Households in Poverty", y = "%POC") 

# how does immigrant status relate to mhi?
ggplot(hh_pov_analysis, aes(x = estimate_hh_below_pov_per, 
                          y = estimate_mhi)) + 
  geom_point() + 
  geom_smooth() + 
  geom_smooth(color = "grey") + 
  theme_minimal() +
  geom_point(data = hh_pov_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  geom_smooth(data = hh_pov_analysis %>% filter(healthbased_violations_5yr>0), 
             color = "red") +
  labs(x = "% Households in Poverty", y = "MHI")
```

## Has the community been identified as disadvantaged by CEJST?
```{r}
cejst <- demo$cejst
pwsids <- demo$census$pwsid
census <- demo$census

# need census tract geographies: 
tract_geo <- tidycensus::get_acs(
  geography = "tract", 
  variables = c(total_pop = "B01003_001"), 
  state = c("TX"), 
  year = 2020,
  geometry = TRUE
)

# CEJST was made off of 2010 tract boundaries - will need to be crosswalked
# for the crosswalk 
tract_2010_2020_crosswalk <- read.table("https://www2.census.gov/geo/docs/maps-data/data/rel2020/tract/tab20_tract20_tract10_natl.txt", 
                                        sep = "|", header = TRUE)

full_crosswalk <- merge(tract_geo, tract_2010_2020_crosswalk, 
                        by.x = "GEOID", 
                        by.y = "GEOID_TRACT_20", all.x = TRUE)

cejst_2020 <- merge(cejst, full_crosswalk, by.x = "census_tract_2010_id", 
                    by.y = "GEOID_TRACT_10", 
                    all.x = T) %>%
  select(GEOID, NAME, 
         total_threshold_criteria_exceeded:percentage_of_tract_that_is_disadvantaged_by_area, 
         geometry)

# using areal weighted interpolation to calculate the % of tract that 
# is disadv by area - should just do areal interpolation
cejst_geo <- cejst_2020 %>%
  st_as_sf() %>%
  st_transform(., crs = "ESRI:102296") 

census_sf <- census %>%
  st_transform(., crs = "ESRI:102296") %>%  
  filter(!st_is_empty(.))%>%
  select(pwsid, geometry)

interpolate <- areal::aw_interpolate(census_sf, tid="pwsid", 
                                     source=cejst_geo, 
                                     sid="GEOID", 
                                     weight = "sum",
                                     output="sf", 
                                     intensive="percentage_of_tract_that_is_disadvantaged_by_area")


# using weighted mean to calculate mean % disadvantaged for pwsids
pwsid_disadv <- interpolate %>%
  group_by(pwsid) %>%
  summarize(per_disadv_mean = mean(percentage_of_tract_that_is_disadvantaged_by_area)) %>%
  as.data.frame() %>%
  select(-geometry)

# combining with key stats of interest: 
key_stats <- keys$analysis_keys
cejst_analysis <- merge(key_stats, pwsid_disadv, by = "pwsid", all.x = TRUE)
cejst_analysis <- cejst_analysis %>%
  filter(east_tx_flag == "yes") %>%
  st_transform(., crs = st_crs(tract_geo))

# plotting by disadv status: 
cejst_pal <- colorNumeric(
  palette = viridis::viridis(9),
  domain = cejst_analysis$per_disadv_mean)

leaflet() %>%
  addProviderTiles(providers$CartoDB.VoyagerNoLabels, group = "Toner Lite") %>%
  addPolygons(data = cejst_analysis,
              fillOpacity = 0.9,
              color = ~cejst_pal(per_disadv_mean),
              weight = 1,
              label = paste0("pwsid: ", cejst_analysis$pwsid, 
                             "; % disadv: ", cejst_analysis$per_disadv_mean)) %>%
  addLegend("bottomright",
            pal = cejst_pal,
            values = cejst_analysis$per_disadv_mean,
            title = "% CEJST Disadv",
            opacity = 1) 

# box plot to investigate each bracket: 
dac_TX <- cejst_analysis %>%
  select(pwsid, per_disadv_mean) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(flag = "TX")
mean(dac_TX$per_disadv_mean, na.rm = T)

dac_eTX <- cejst_analysis %>%
  select(pwsid, east_tx_flag, per_disadv_mean) %>%
  as.data.frame() %>%
  select(-geometry) %>%
  filter(east_tx_flag == "yes") %>%
  select(-east_tx_flag) %>%
  mutate(flag = "East TX")
mean(dac_eTX$per_disadv_mean, na.rm = T)

all_dac <- rbind(dac_TX, dac_eTX)

# running boxplot: 
ggplot(all_dac, aes(x = flag,  
                    y = per_disadv_mean, 
                    fill = flag)) + 
  geom_violin(alpha = 0.5) + 
  scale_fill_discrete(name = "") +
  theme_minimal()  + 
  labs(x = "", y = "Percent of Service Area Disadv by CEJST") 

# looking at data by quantile of disadv status: 
cejst_analysis %>%
  as.data.frame() %>%
  select(-geometry) %>%
  mutate(per_disadv_mean = as.numeric(per_disadv_mean)) %>%
  mutate(dac_bin = cut(per_disadv_mean, 
                       breaks = c(unique(quantile(per_disadv_mean, 
                                           probs = seq(0, 1, by = 0.05), 
                                           na.rm = TRUE))), 
                       include.lowest = TRUE, 
                       dig.lab = 18)) %>%
  group_by(dac_bin) %>%
  summarize(total_utilities = n(),
            paperwork_10yr = mean(paperwork_violations_10yr), 
            paperwork_5yr = mean(paperwork_violations_5yr), 
            health_10yr = mean(healthbased_violations_10yr), 
            health_5yr = mean(healthbased_violations_5yr), 
            mean_mhi = mean(estimate_mhi, na.rm = T), 
            per_poc  = mean(estimate_poc_alone_per, na.rm = T)) 

# what does this look like in the form of a graph?
ggplot(cejst_analysis, aes(x = per_disadv_mean, 
                           y = estimate_mhi, color = healthbased_violations_5yr)) + 
  geom_point() + 
  theme_minimal() + 
  labs(x = "% of SAB that is Disadvantaged", y = "Median Household Income") + 
  scale_color_continuous(name = "Heath-based Violations (5yr)") + 
  theme(legend.position = "bottom")

# are there any clear correlations? 
all_cor <- cor(cejst_analysis %>% 
                 as.data.frame() %>%
                 select(-east_tx_flag) %>%
                 select(c(healthbased_violations_5yr,
                          paperwork_violations_5yr, 
                          paperwork_violations_10yr,
                          healthbased_violations_10yr,
                          estimate_mhi, 
                          estimate_poc_alone_per, 
                          per_disadv_mean)), 
               use = "pairwise.complete.obs")

cor_pvals <- cor.mtest(all_cor, conf.level = 0.95)
cor_pvals$p

corrplot(all_cor, method = 'ellipse', type = 'upper', 
        p.mat = cor_pvals$p, insig = "p-value", sig.level = -1, 
        mar=c(0,0,0,0))


# can we link to which vulnerabilities are being exceeded?
# Just grabbing counties in East TX
east_tx <- read.csv("./data/raw/east_TX_counties.csv") %>%
  select(-X) %>%
  mutate(county_tidy = paste0(county_name, " County"))

cejst_raw <- aws.s3::s3read_using(read.csv, 
                                  object = "s3://tech-team-data/ej-data/cejest-v1/1.0-communities.csv") %>%
  janitor::clean_names() %>%
  filter(state_territory == "Texas") %>%
  filter(county_name %in% east_tx$county_tidy)

# grabbing mean percentile for each county: 
percentiles <- cejst_raw[, endsWith(names(cejst_raw), '_percentile')]
id_cols <- cejst_raw[,c("census_tract_2010_id", "county_name")]
cejst_percentiles <- cbind(id_cols, percentiles)
county_percentiles <- cejst_percentiles %>%
  group_by(county_name) %>%
  summarise(across(!c("census_tract_2010_id"), ~mean(.x, na.rm = T)))

etx_percentiles <- county_percentiles %>%
  summarise(across(!c("county_name"), ~mean(.x, na.rm = T))) %>%
  pivot_longer(1:31)
```

## Investigating stats for cws in TX and comparisons across different scales:
```{r}
# grabbing stats at different scales: 
census <- demo$census
east_tx_cws <- census %>%
  filter(east_tx_flag == "yes")

sum(census$estimate_total_pop, na.rm = T)
sum(east_tx_cws$estimate_total_pop, na.rm = T)

# summaries for those served by community water systems: 
summary(census)
summary(east_tx_cws)

################################################################################
# How does this compare to all of Texas?
################################################################################
census_vars <- c(total_pop = "B01003_001",
                 # race and ethnicity stats: 
                 black_alone = "B02001_003", 
                 asian_alone = "B02001_005", 
                 white_alone = "B02001_002", 
                 AIAN_alone = "B02001_004", 
                 NAPI_alone = "B02001_006", 
                 other_alone = "B02001_007", 
                 mixed_alone = "B02001_008",
                 hisp_alone = "B03003_003", 
                 # MHI
                 mhi = "B19013_001", 
                 # labor force and unemployment: 
                 laborforce = "B23025_003",  # universe: pop > 16
                 laborforce_unemployed = "B23025_005", # universe: pop > 16
                 # households below poverty level: 
                 hh_total = "B17017_001",
                 hh_below_pov = "B17017_002", 
                 # age stats: 
                 ageunder_5  = "B06001_002",
                 age5_17 = "B06001_003", 
                 age18_24 = "B06001_004",
                 age25_34 = "B06001_005", 
                 age35_44 = "B06001_006", 
                 age45_54 = "B06001_007", 
                 age55_59 = "B06001_008", 
                 age60_61 = "B06001_009",
                 # language spoken at home: 
                 only_english = "B99162_002",  # universe: pop >5
                 other_lang = "B99162_003", # universe: pop >5
                 # education categories: 
                 no_school = "B15003_002", 
                 high_school = "B15003_017", 
                 bachelors = "B15003_022",
                 prof_degree = "B15003_024", 
                 # nationality: 
                 foreign = "B99051_005")
tx_census <- tidycensus::get_acs(
  geography = "state", 
  variables = census_vars, 
  state = c("TX"), 
  year = 2020,
  geometry = TRUE
)

# stats for all of TX: 
tx_wide <- pivot_wider(tx_census, 
                       names_from = c("variable"), 
                       values_from = c("estimate", "moe"))

tx_wide <- tx_wide %>%
  select(starts_with(c("estimate"))) %>%
  # calculating some columns where the denominator is not the total pop: 
  mutate(estimate_laborforce_unemployed = ((estimate_laborforce_unemployed/estimate_laborforce)*100), 
         estimate_hh_below_pov = ((estimate_hh_below_pov/estimate_hh_total)*100)) %>%
  # leaving out columns that are totals or mhi:
  mutate_at(vars(!c("estimate_total_pop", "estimate_mhi", 
                    "estimate_laborforce_unemployed", "estimate_hh_below_pov")), 
            ~((./estimate_total_pop)*100)) %>%
  mutate(estimate_poc_alone = (100 - estimate_white_alone))

summary(tx_wide)

################################################################################
# How does this compare to all of East Texas?
################################################################################
# grabbing East TX stats: 
etx_census <- tidycensus::get_acs(
  # geography = "county", 
  geography = "tract", 
  variables = census_vars, 
  state = c("TX"),
  year = 2020,
  geometry = TRUE
)

counties <- unlist(strsplit(etx_census$NAME, split = ","))
etx_census$counties <- trimws(counties[grepl("County", counties)])

east_tx <- read.csv("./data/raw/east_TX_counties.csv") %>%
  select(-X) %>%
  mutate(county_tidy = paste0(county_name, " County"))

east_tx_geo <- etx_census %>%
  filter(counties %in% east_tx$county_tidy) 

etx_wide <- pivot_wider(east_tx_geo, 
                       names_from = c("variable"), 
                       values_from = c("estimate", "moe"))

etx_wide_per <- etx_wide %>%
  select(starts_with(c("estimate"))) %>%
  # calculating some columns where the denominator is not the total pop: 
  mutate(estimate_laborforce_unemployed = ((estimate_laborforce_unemployed/estimate_laborforce)*100), 
         estimate_hh_below_pov = ((estimate_hh_below_pov/estimate_hh_total)*100)) %>%
  # leaving out columns that are totals or mhi:
  mutate_at(vars(!c("estimate_total_pop", "estimate_mhi", 
                    "estimate_laborforce_unemployed", 
                    "estimate_hh_below_pov")), 
            ~((./estimate_total_pop)*100)) %>%
  mutate(estimate_poc_alone = (100 - estimate_white_alone))

summary(etx_wide_per)
sum(etx_wide$estimate_total_pop)
```

## Where are POC located at the tract level?
```{r}
etx_wide_sf <- etx_wide %>%
  mutate(percent_poc = (100*(estimate_total_pop - estimate_white_alone)/estimate_total_pop))

poc_pal <- colorNumeric(
  palette = viridis::viridis(9),
  domain = etx_wide_sf$percent_poc)

leaflet() %>%
    addPolygons(data = etx_wide_sf,
              fillOpacity = 0.3,
              color = ~poc_pal(percent_poc),
              weight = 1,
              label = paste0("% poc: ", etx_wide_sf$percent_poc)) %>%
  addProviderTiles(providers$CartoDB.VoyagerLabels, group = "Toner Lite") %>%
  addLegend("bottomright",
            pal = poc_pal,
            values = etx_wide_sf$percent_poc,
            title = "% POC",
            opacity = 1) 
```

## Correlations between variables: 
```{r}
census <- demo$census 

# correlation plot across all metrics: 
all_data <- census %>%
  as.data.frame() %>%
  left_join(keys$analysis_keys) %>%
  # focusing on East TX for now: 
  filter(east_tx_flag == "yes")

all_cor <- cor(all_data %>% 
                 select(c(healthbased_violations_5yr,
                          paperwork_violations_5yr, 
                          healthbased_violations_10yr,
                          paperwork_violations_10yr,
                          area_miles, pop_density, 
                          service_connections_count,
                          estimate_mhi, 
                          estimate_hh_below_pov_per,
                          estimate_laborforce_unemployed, 
                          estimate_no_school_per, 
                          estimate_high_school_per,
                          estimate_only_english_per,
                          estimate_foreign_per,
                          estimate_hisp_alone_per,
                          estimate_poc_alone_per)) %>% 
                 rename_with(~ str_remove(., "_alone_per"),
                             everything()) %>%
                          rename_with(~ str_remove(., "_per"), 
                                      everything()) %>%
                          rename_with(~ str_remove(., "estimate"),
                                      everything()) %>%
                          rename_with(~ str_remove(., "_"), 
                                      everything()), 
               use = "pairwise.complete.obs") 


name_changes <- c("Health Viols, 5yr", "Non-Health Viols, 5yr", 
                      "Health Viols, 10yr",  "Non-Helath Viols, 10yr", 
                       "Area (mi)", "Population Density", 
                       "Service Connections", 
                       "MHI", "% Households in Poverty", 
                       "% Unemployment", "% No Schooling", 
                       "% Highschool", "% Only Speak English",
                       "% Immigrant", "% Latino/a", "% POC")

rownames(all_cor) <- name_changes
colnames(all_cor) <- name_changes

cor_pvals <- cor.mtest(all_cor, conf.level = 0.95)
cor_pvals$p

corrplot(all_cor, method = 'ellipse', type = 'upper', 
         p.mat = cor_pvals$p, insig = "p-value", sig.level = -1, 
         mar=c(0,0,0,0), 
         tl.col = "black", 
         number.cex = 0.2)
```
